Move high_low_state building code to diff file
