+++ rubikscubennnsolver/RubiksCube444.py
@@ -733,185 +733,6 @@ class RubiksCube444(RubiksCube):
         self._sanity_check('centers', centers_444, 4)
         self._sanity_check('edge-orbit-0', edge_orbit_0, 8)
 
-    def high_low_state(self, x, y, state_x, state_y, wing_str):
-        """
-        Return U if this is a high edge, D if it is a low edge
-        """
-        original_state = self.state[:]
-        original_solution = self.solution[:]
-
-        # Nuke everything except the one wing we are interested in
-        self.nuke_corners()
-        self.nuke_centers()
-        self.nuke_edges()
-
-        self.state[x] = state_x
-        self.state[y] = state_y
-
-        # Now move that wing to its home edge
-        if wing_str.startswith('U'):
-
-            if wing_str == 'UB':
-                self.move_wing_to_U_north(x)
-                high_edge_index = 2
-                low_edge_index = 3
-
-            elif wing_str == 'UL':
-                self.move_wing_to_U_west(x)
-                high_edge_index = 9
-                low_edge_index = 5
-
-            elif wing_str == 'UR':
-                self.move_wing_to_U_east(x)
-                high_edge_index = 8
-                low_edge_index = 12
-
-            elif wing_str == 'UF':
-                self.move_wing_to_U_south(x)
-                high_edge_index = 15
-                low_edge_index = 14
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if self.state[high_edge_index] == 'U':
-                result = 'U'
-            elif self.state[low_edge_index] == 'U':
-                result = 'D'
-            elif self.state[high_edge_index] == 'x':
-                result = 'U'
-            elif self.state[low_edge_index] == 'x':
-                result = 'D'
-            else:
-                self.print_cube()
-                raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                    (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        elif wing_str.startswith('L'):
-
-            if wing_str == 'LB':
-                self.move_wing_to_L_west(x)
-                high_edge_index = 25
-                low_edge_index = 21
-
-            elif wing_str == 'LF':
-                self.move_wing_to_L_east(x)
-                high_edge_index = 24
-                low_edge_index = 28
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if self.state[high_edge_index] == 'L':
-                result = 'U'
-            elif self.state[low_edge_index] == 'L':
-                result = 'D'
-            elif self.state[high_edge_index] == 'x':
-                result = 'U'
-            elif self.state[low_edge_index] == 'x':
-                result = 'D'
-            else:
-                self.print_cube()
-                raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                    (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        elif wing_str.startswith('R'):
-
-            if wing_str == 'RB':
-                self.move_wing_to_R_east(x)
-                high_edge_index = 56
-                low_edge_index = 60
-
-            elif wing_str == 'RF':
-                self.move_wing_to_R_west(x)
-                high_edge_index = 57
-                low_edge_index = 53
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if self.state[high_edge_index] == 'R':
-                result = 'U'
-            elif self.state[low_edge_index] == 'R':
-                result = 'D'
-            elif self.state[high_edge_index] == 'x':
-                result = 'U'
-            elif self.state[low_edge_index] == 'x':
-                result = 'D'
-            else:
-                self.print_cube()
-                raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                    (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        elif wing_str.startswith('D'):
-            if wing_str == 'DB':
-                self.move_wing_to_D_south(x)
-                high_edge_index = 95
-                low_edge_index = 94
-
-            elif wing_str == 'DL':
-                self.move_wing_to_D_west(x)
-                high_edge_index = 89
-                low_edge_index = 85
-
-            elif wing_str == 'DR':
-                self.move_wing_to_D_east(x)
-                high_edge_index = 88
-                low_edge_index = 92
-
-            elif wing_str == 'DF':
-                self.move_wing_to_D_north(x)
-                high_edge_index = 82
-                low_edge_index = 83
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if self.state[high_edge_index] == 'D':
-                result = 'U'
-            elif self.state[low_edge_index] == 'D':
-                result = 'D'
-            elif self.state[high_edge_index] == 'x':
-                result = 'U'
-            elif self.state[low_edge_index] == 'x':
-                result = 'D'
-            else:
-                self.print_cube()
-                raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                    (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        else:
-            raise Exception("invalid wing_str %s" % wing_str)
-
-        self.state = original_state[:]
-        self.solution = original_solution[:]
-
-        assert result in ('U', 'D')
-        return result
-
-    def build_highlow_edge_values(self):
-        state = self.state
-        new_highlow_edge_values = {}
-
-        for x in range(1000000):
-
-            # make random moves
-            step = moves_444[randint(0, len(moves_444)-1)]
-            self.rotate(step)
-
-            for (x, y) in self.reduce333_orient_edges_tuples:
-                state_x = self.state[x]
-                state_y = self.state[y]
-                wing_str = wing_str_map[state_x + state_y]
-                wing_tuple = (x, y, state_x, state_y)
-
-                if wing_tuple not in new_highlow_edge_values:
-                    new_highlow_edge_values[wing_tuple] = self.high_low_state(x, y, state_x, state_y, wing_str)
-
-        print("new highlow_edge_values\n\n%s\n\n" % pformat(new_highlow_edge_values))
-        log.info("new_highlow_edge_values has %d entries" % len(new_highlow_edge_values))
-        sys.exit(0)
-
     def highlow_edges_state(self, edges_to_flip):
         state = self.state
