+++ rubikscubennnsolver/RubiksCube444Misc.py
@@ -29,7 +29,6 @@ low_edges_444 = ((2, 3, 66),  # upper
                  (7, 92, 63),
                  (4, 94, 79))
 
-
 highlow_edge_values = {
     (2, 67, 'B', 'D'): 'D',
     (2, 67, 'B', 'L'): 'D',
