+++ rubikscubennnsolver/RubiksCube555.py
@@ -4,6 +4,7 @@ from rubikscubennnsolver import RubiksCube, NotSolving, wing_str_map, wing_strs_
 from rubikscubennnsolver.misc import pre_steps_to_try, pre_steps_stage_l4e, wing_str_combos_two, wing_str_combos_four
 from rubikscubennnsolver.RubiksSide import SolveError
 from rubikscubennnsolver.RubiksCube444 import RubiksCube444, solved_444
+from rubikscubennnsolver.RubiksCube555Misc import highlow_edge_values
 from rubikscubennnsolver.LookupTable import (
     steps_on_same_face_and_layer,
     LookupTable,
@@ -1773,302 +1774,6 @@ class RubiksCube555(RubiksCube):
         self.lt_pair_last_eight_edges_centers_only = LookupTable555PairLastEightEdgesCentersOnly(self)
         self.lt_pair_last_eight_edges_centers_only.preload_cache_dict()
 
-    def high_low_state(self, x, y, state_x, state_y, wing_str):
-        """
-        Return U if this is a high edge, D if it is a low edge
-        """
-        original_state = self.state[:]
-        original_solution = self.solution[:]
-
-        # Nuke everything except the one wing we are interested in
-        self.nuke_corners()
-        self.nuke_centers()
-        self.nuke_edges()
-
-        self.state[x] = state_x
-        self.state[y] = state_y
-
-        if x in midge_indexes or y in midge_indexes:
-            is_midge = True
-        else:
-            is_midge = False
-
-        # Now move that wing to its home edge
-        if wing_str.startswith('U'):
-
-            if wing_str == 'UB':
-                self.move_wing_to_U_north(x)
-
-                if is_midge:
-                    high_edge_index = 3
-                    low_edge_index = 103
-                else:
-                    high_edge_index = 2
-                    low_edge_index = 4
-
-            elif wing_str == 'UL':
-                self.move_wing_to_U_west(x)
-
-                if is_midge:
-                    high_edge_index = 11
-                    low_edge_index = 28
-                else:
-                    high_edge_index = 16
-                    low_edge_index = 6
-
-            elif wing_str == 'UR':
-                self.move_wing_to_U_east(x)
-
-                if is_midge:
-                    high_edge_index = 15
-                    low_edge_index = 78
-                else:
-                    high_edge_index = 10
-                    low_edge_index = 20
-
-            elif wing_str == 'UF':
-                self.move_wing_to_U_south(x)
-
-                if is_midge:
-                    high_edge_index = 23
-                    low_edge_index = 53
-                else:
-                    high_edge_index = 24
-                    low_edge_index = 22
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if is_midge:
-                if self.state[high_edge_index] == 'U':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'U':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s, is_midge %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index], is_midge))
-            else:
-                if self.state[high_edge_index] == 'U':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'U':
-                    result = 'D'
-                elif self.state[high_edge_index] == '.':
-                    result = 'U'
-                elif self.state[low_edge_index] == '.':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s, is_midge %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index], is_midge))
-
-        elif wing_str.startswith('L'):
-
-            if wing_str == 'LB':
-                self.move_wing_to_L_west(x)
-
-                if is_midge:
-                    high_edge_index = 36
-                    low_edge_index = 115
-                else:
-                    high_edge_index = 41
-                    low_edge_index = 31
-
-            elif wing_str == 'LF':
-                self.move_wing_to_L_east(x)
-
-                if is_midge:
-                    high_edge_index = 40
-                    low_edge_index = 61
-                else:
-                    high_edge_index = 35
-                    low_edge_index = 45
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if is_midge:
-                if self.state[high_edge_index] == 'L':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'L':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s, is_midge %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index], is_midge))
-            else:
-                if self.state[high_edge_index] == 'L':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'L':
-                    result = 'D'
-                elif self.state[high_edge_index] == '.':
-                    result = 'U'
-                elif self.state[low_edge_index] == '.':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        elif wing_str.startswith('R'):
-
-            if wing_str == 'RB':
-                self.move_wing_to_R_east(x)
-
-                if is_midge:
-                    high_edge_index = 90
-                    low_edge_index = 111
-                else:
-                    high_edge_index = 85
-                    low_edge_index = 95
-
-            elif wing_str == 'RF':
-                self.move_wing_to_R_west(x)
-
-                if is_midge:
-                    high_edge_index = 86
-                    low_edge_index = 65
-                else:
-                    high_edge_index = 91
-                    low_edge_index = 81
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if is_midge:
-                if self.state[high_edge_index] == 'R':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'R':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s, is_midge %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index], is_midge))
-            else:
-                if self.state[high_edge_index] == 'R':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'R':
-                    result = 'D'
-                elif self.state[high_edge_index] == '.':
-                    result = 'U'
-                elif self.state[low_edge_index] == '.':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        elif wing_str.startswith('D'):
-            if wing_str == 'DB':
-                self.move_wing_to_D_south(x)
-
-                if is_midge:
-                    high_edge_index = 148
-                    low_edge_index = 123
-                else:
-                    high_edge_index = 149
-                    low_edge_index = 147
-
-            elif wing_str == 'DL':
-                self.move_wing_to_D_west(x)
-
-                if is_midge:
-                    high_edge_index = 136
-                    low_edge_index = 48
-                else:
-                    high_edge_index = 141
-                    low_edge_index = 131
-
-            elif wing_str == 'DR':
-                self.move_wing_to_D_east(x)
-
-                if is_midge:
-                    high_edge_index = 140
-                    low_edge_index = 98
-                else:
-                    high_edge_index = 135
-                    low_edge_index = 145
-
-            elif wing_str == 'DF':
-                self.move_wing_to_D_north(x)
-
-                if is_midge:
-                    high_edge_index = 128
-                    low_edge_index = 73
-                else:
-                    high_edge_index = 127
-                    low_edge_index = 129
-
-            else:
-                raise Exception("invalid wing_str %s" % wing_str)
-
-            if is_midge:
-                if self.state[high_edge_index] == 'D':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'D':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s, is_midge %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index], is_midge))
-            else:
-                if self.state[high_edge_index] == 'D':
-                    result = 'U'
-                elif self.state[low_edge_index] == 'D':
-                    result = 'D'
-                elif self.state[high_edge_index] == '.':
-                    result = 'U'
-                elif self.state[low_edge_index] == '.':
-                    result = 'D'
-                else:
-                    self.print_cube()
-                    self.print_cube_layout()
-                    raise Exception("something went wrong, (%s, %s) was originally (%s, %s), moved to %s, high_index state[%s] is %s, low_index state[%s] is %s" %
-                        (x, y, state_x, state_y, wing_str, high_edge_index, self.state[high_edge_index], low_edge_index, self.state[low_edge_index]))
-
-        else:
-            raise Exception("invalid wing_str %s" % wing_str)
-
-        self.state = original_state[:]
-        self.solution = original_solution[:]
-
-        assert result in ('U', 'D')
-        return result
-
-    def build_highlow_edge_values(self):
-        state = self.state
-        new_highlow_edge_values = {}
-
-        for x in range(1000000):
-
-            # make random moves
-            step = moves_555[randint(0, len(moves_555)-1)]
-
-            if "w" in step:
-                continue
-
-            self.rotate(step)
-
-            for (x, y) in self.reduce333_orient_edges_tuples:
-                state_x = self.state[x]
-                state_y = self.state[y]
-                wing_str = wing_str_map[state_x + state_y]
-                wing_tuple = (x, y, state_x, state_y)
-
-                if wing_tuple not in new_highlow_edge_values:
-                    new_highlow_edge_values[wing_tuple] = self.high_low_state(x, y, state_x, state_y, wing_str)
-
-        print("new highlow_edge_values\n\n%s\n\n" % pformat(new_highlow_edge_values))
-        log.info("new_highlow_edge_values has %d entries" % len(new_highlow_edge_values))
-        sys.exit(0)
-
     def highlow_edges_state_x_plane_y_plane(self):
         state = self.state
         result = [highlow_edge_values[(x, y, state[x], state[y])] for (x, y) in self.reduce333_orient_edges_x_plane_y_plane_tuples]
@@ -2368,1738 +2073,6 @@ class RubiksCube555(RubiksCube):
 
         if not fake_555:
             self.solution.append('EDGES_GROUPED')
-
-
-highlow_edge_values = {
-    (2, 104, 'B', 'D'): 'D',
-    (2, 104, 'B', 'L'): 'D',
-    (2, 104, 'B', 'R'): 'D',
-    (2, 104, 'B', 'U'): 'D',
-    (2, 104, 'D', 'B'): 'U',
-    (2, 104, 'D', 'F'): 'U',
-    (2, 104, 'D', 'L'): 'U',
-    (2, 104, 'D', 'R'): 'U',
-    (2, 104, 'F', 'D'): 'D',
-    (2, 104, 'F', 'L'): 'D',
-    (2, 104, 'F', 'R'): 'D',
-    (2, 104, 'F', 'U'): 'D',
-    (2, 104, 'L', 'B'): 'U',
-    (2, 104, 'L', 'D'): 'D',
-    (2, 104, 'L', 'F'): 'U',
-    (2, 104, 'L', 'U'): 'D',
-    (2, 104, 'R', 'B'): 'U',
-    (2, 104, 'R', 'D'): 'D',
-    (2, 104, 'R', 'F'): 'U',
-    (2, 104, 'R', 'U'): 'D',
-    (2, 104, 'U', 'B'): 'U',
-    (2, 104, 'U', 'F'): 'U',
-    (2, 104, 'U', 'L'): 'U',
-    (2, 104, 'U', 'R'): 'U',
-    (3, 103, 'B', 'D'): 'D',
-    (3, 103, 'B', 'L'): 'D',
-    (3, 103, 'B', 'R'): 'D',
-    (3, 103, 'B', 'U'): 'D',
-    (3, 103, 'D', 'B'): 'U',
-    (3, 103, 'D', 'F'): 'U',
-    (3, 103, 'D', 'L'): 'U',
-    (3, 103, 'D', 'R'): 'U',
-    (3, 103, 'F', 'D'): 'D',
-    (3, 103, 'F', 'L'): 'D',
-    (3, 103, 'F', 'R'): 'D',
-    (3, 103, 'F', 'U'): 'D',
-    (3, 103, 'L', 'B'): 'U',
-    (3, 103, 'L', 'D'): 'D',
-    (3, 103, 'L', 'F'): 'U',
-    (3, 103, 'L', 'U'): 'D',
-    (3, 103, 'R', 'B'): 'U',
-    (3, 103, 'R', 'D'): 'D',
-    (3, 103, 'R', 'F'): 'U',
-    (3, 103, 'R', 'U'): 'D',
-    (3, 103, 'U', 'B'): 'U',
-    (3, 103, 'U', 'F'): 'U',
-    (3, 103, 'U', 'L'): 'U',
-    (3, 103, 'U', 'R'): 'U',
-    (4, 102, 'B', 'D'): 'U',
-    (4, 102, 'B', 'L'): 'U',
-    (4, 102, 'B', 'R'): 'U',
-    (4, 102, 'B', 'U'): 'U',
-    (4, 102, 'D', 'B'): 'D',
-    (4, 102, 'D', 'F'): 'D',
-    (4, 102, 'D', 'L'): 'D',
-    (4, 102, 'D', 'R'): 'D',
-    (4, 102, 'F', 'D'): 'U',
-    (4, 102, 'F', 'L'): 'U',
-    (4, 102, 'F', 'R'): 'U',
-    (4, 102, 'F', 'U'): 'U',
-    (4, 102, 'L', 'B'): 'D',
-    (4, 102, 'L', 'D'): 'U',
-    (4, 102, 'L', 'F'): 'D',
-    (4, 102, 'L', 'U'): 'U',
-    (4, 102, 'R', 'B'): 'D',
-    (4, 102, 'R', 'D'): 'U',
-    (4, 102, 'R', 'F'): 'D',
-    (4, 102, 'R', 'U'): 'U',
-    (4, 102, 'U', 'B'): 'D',
-    (4, 102, 'U', 'F'): 'D',
-    (4, 102, 'U', 'L'): 'D',
-    (4, 102, 'U', 'R'): 'D',
-    (6, 27, 'B', 'D'): 'U',
-    (6, 27, 'B', 'L'): 'U',
-    (6, 27, 'B', 'R'): 'U',
-    (6, 27, 'B', 'U'): 'U',
-    (6, 27, 'D', 'B'): 'D',
-    (6, 27, 'D', 'F'): 'D',
-    (6, 27, 'D', 'L'): 'D',
-    (6, 27, 'D', 'R'): 'D',
-    (6, 27, 'F', 'D'): 'U',
-    (6, 27, 'F', 'L'): 'U',
-    (6, 27, 'F', 'R'): 'U',
-    (6, 27, 'F', 'U'): 'U',
-    (6, 27, 'L', 'B'): 'D',
-    (6, 27, 'L', 'D'): 'U',
-    (6, 27, 'L', 'F'): 'D',
-    (6, 27, 'L', 'U'): 'U',
-    (6, 27, 'R', 'B'): 'D',
-    (6, 27, 'R', 'D'): 'U',
-    (6, 27, 'R', 'F'): 'D',
-    (6, 27, 'R', 'U'): 'U',
-    (6, 27, 'U', 'B'): 'D',
-    (6, 27, 'U', 'F'): 'D',
-    (6, 27, 'U', 'L'): 'D',
-    (6, 27, 'U', 'R'): 'D',
-    (10, 79, 'B', 'D'): 'D',
-    (10, 79, 'B', 'L'): 'D',
-    (10, 79, 'B', 'R'): 'D',
-    (10, 79, 'B', 'U'): 'D',
-    (10, 79, 'D', 'B'): 'U',
-    (10, 79, 'D', 'F'): 'U',
-    (10, 79, 'D', 'L'): 'U',
-    (10, 79, 'D', 'R'): 'U',
-    (10, 79, 'F', 'D'): 'D',
-    (10, 79, 'F', 'L'): 'D',
-    (10, 79, 'F', 'R'): 'D',
-    (10, 79, 'F', 'U'): 'D',
-    (10, 79, 'L', 'B'): 'U',
-    (10, 79, 'L', 'D'): 'D',
-    (10, 79, 'L', 'F'): 'U',
-    (10, 79, 'L', 'U'): 'D',
-    (10, 79, 'R', 'B'): 'U',
-    (10, 79, 'R', 'D'): 'D',
-    (10, 79, 'R', 'F'): 'U',
-    (10, 79, 'R', 'U'): 'D',
-    (10, 79, 'U', 'B'): 'U',
-    (10, 79, 'U', 'F'): 'U',
-    (10, 79, 'U', 'L'): 'U',
-    (10, 79, 'U', 'R'): 'U',
-    (11, 28, 'B', 'D'): 'D',
-    (11, 28, 'B', 'L'): 'D',
-    (11, 28, 'B', 'R'): 'D',
-    (11, 28, 'B', 'U'): 'D',
-    (11, 28, 'D', 'B'): 'U',
-    (11, 28, 'D', 'F'): 'U',
-    (11, 28, 'D', 'L'): 'U',
-    (11, 28, 'D', 'R'): 'U',
-    (11, 28, 'F', 'D'): 'D',
-    (11, 28, 'F', 'L'): 'D',
-    (11, 28, 'F', 'R'): 'D',
-    (11, 28, 'F', 'U'): 'D',
-    (11, 28, 'L', 'B'): 'U',
-    (11, 28, 'L', 'D'): 'D',
-    (11, 28, 'L', 'F'): 'U',
-    (11, 28, 'L', 'U'): 'D',
-    (11, 28, 'R', 'B'): 'U',
-    (11, 28, 'R', 'D'): 'D',
-    (11, 28, 'R', 'F'): 'U',
-    (11, 28, 'R', 'U'): 'D',
-    (11, 28, 'U', 'B'): 'U',
-    (11, 28, 'U', 'F'): 'U',
-    (11, 28, 'U', 'L'): 'U',
-    (11, 28, 'U', 'R'): 'U',
-    (15, 78, 'B', 'D'): 'D',
-    (15, 78, 'B', 'L'): 'D',
-    (15, 78, 'B', 'R'): 'D',
-    (15, 78, 'B', 'U'): 'D',
-    (15, 78, 'D', 'B'): 'U',
-    (15, 78, 'D', 'F'): 'U',
-    (15, 78, 'D', 'L'): 'U',
-    (15, 78, 'D', 'R'): 'U',
-    (15, 78, 'F', 'D'): 'D',
-    (15, 78, 'F', 'L'): 'D',
-    (15, 78, 'F', 'R'): 'D',
-    (15, 78, 'F', 'U'): 'D',
-    (15, 78, 'L', 'B'): 'U',
-    (15, 78, 'L', 'D'): 'D',
-    (15, 78, 'L', 'F'): 'U',
-    (15, 78, 'L', 'U'): 'D',
-    (15, 78, 'R', 'B'): 'U',
-    (15, 78, 'R', 'D'): 'D',
-    (15, 78, 'R', 'F'): 'U',
-    (15, 78, 'R', 'U'): 'D',
-    (15, 78, 'U', 'B'): 'U',
-    (15, 78, 'U', 'F'): 'U',
-    (15, 78, 'U', 'L'): 'U',
-    (15, 78, 'U', 'R'): 'U',
-    (16, 29, 'B', 'D'): 'D',
-    (16, 29, 'B', 'L'): 'D',
-    (16, 29, 'B', 'R'): 'D',
-    (16, 29, 'B', 'U'): 'D',
-    (16, 29, 'D', 'B'): 'U',
-    (16, 29, 'D', 'F'): 'U',
-    (16, 29, 'D', 'L'): 'U',
-    (16, 29, 'D', 'R'): 'U',
-    (16, 29, 'F', 'D'): 'D',
-    (16, 29, 'F', 'L'): 'D',
-    (16, 29, 'F', 'R'): 'D',
-    (16, 29, 'F', 'U'): 'D',
-    (16, 29, 'L', 'B'): 'U',
-    (16, 29, 'L', 'D'): 'D',
-    (16, 29, 'L', 'F'): 'U',
-    (16, 29, 'L', 'U'): 'D',
-    (16, 29, 'R', 'B'): 'U',
-    (16, 29, 'R', 'D'): 'D',
-    (16, 29, 'R', 'F'): 'U',
-    (16, 29, 'R', 'U'): 'D',
-    (16, 29, 'U', 'B'): 'U',
-    (16, 29, 'U', 'F'): 'U',
-    (16, 29, 'U', 'L'): 'U',
-    (16, 29, 'U', 'R'): 'U',
-    (20, 77, 'B', 'D'): 'U',
-    (20, 77, 'B', 'L'): 'U',
-    (20, 77, 'B', 'R'): 'U',
-    (20, 77, 'B', 'U'): 'U',
-    (20, 77, 'D', 'B'): 'D',
-    (20, 77, 'D', 'F'): 'D',
-    (20, 77, 'D', 'L'): 'D',
-    (20, 77, 'D', 'R'): 'D',
-    (20, 77, 'F', 'D'): 'U',
-    (20, 77, 'F', 'L'): 'U',
-    (20, 77, 'F', 'R'): 'U',
-    (20, 77, 'F', 'U'): 'U',
-    (20, 77, 'L', 'B'): 'D',
-    (20, 77, 'L', 'D'): 'U',
-    (20, 77, 'L', 'F'): 'D',
-    (20, 77, 'L', 'U'): 'U',
-    (20, 77, 'R', 'B'): 'D',
-    (20, 77, 'R', 'D'): 'U',
-    (20, 77, 'R', 'F'): 'D',
-    (20, 77, 'R', 'U'): 'U',
-    (20, 77, 'U', 'B'): 'D',
-    (20, 77, 'U', 'F'): 'D',
-    (20, 77, 'U', 'L'): 'D',
-    (20, 77, 'U', 'R'): 'D',
-    (22, 52, 'B', 'D'): 'U',
-    (22, 52, 'B', 'L'): 'U',
-    (22, 52, 'B', 'R'): 'U',
-    (22, 52, 'B', 'U'): 'U',
-    (22, 52, 'D', 'B'): 'D',
-    (22, 52, 'D', 'F'): 'D',
-    (22, 52, 'D', 'L'): 'D',
-    (22, 52, 'D', 'R'): 'D',
-    (22, 52, 'F', 'D'): 'U',
-    (22, 52, 'F', 'L'): 'U',
-    (22, 52, 'F', 'R'): 'U',
-    (22, 52, 'F', 'U'): 'U',
-    (22, 52, 'L', 'B'): 'D',
-    (22, 52, 'L', 'D'): 'U',
-    (22, 52, 'L', 'F'): 'D',
-    (22, 52, 'L', 'U'): 'U',
-    (22, 52, 'R', 'B'): 'D',
-    (22, 52, 'R', 'D'): 'U',
-    (22, 52, 'R', 'F'): 'D',
-    (22, 52, 'R', 'U'): 'U',
-    (22, 52, 'U', 'B'): 'D',
-    (22, 52, 'U', 'F'): 'D',
-    (22, 52, 'U', 'L'): 'D',
-    (22, 52, 'U', 'R'): 'D',
-    (23, 53, 'B', 'D'): 'D',
-    (23, 53, 'B', 'L'): 'D',
-    (23, 53, 'B', 'R'): 'D',
-    (23, 53, 'B', 'U'): 'D',
-    (23, 53, 'D', 'B'): 'U',
-    (23, 53, 'D', 'F'): 'U',
-    (23, 53, 'D', 'L'): 'U',
-    (23, 53, 'D', 'R'): 'U',
-    (23, 53, 'F', 'D'): 'D',
-    (23, 53, 'F', 'L'): 'D',
-    (23, 53, 'F', 'R'): 'D',
-    (23, 53, 'F', 'U'): 'D',
-    (23, 53, 'L', 'B'): 'U',
-    (23, 53, 'L', 'D'): 'D',
-    (23, 53, 'L', 'F'): 'U',
-    (23, 53, 'L', 'U'): 'D',
-    (23, 53, 'R', 'B'): 'U',
-    (23, 53, 'R', 'D'): 'D',
-    (23, 53, 'R', 'F'): 'U',
-    (23, 53, 'R', 'U'): 'D',
-    (23, 53, 'U', 'B'): 'U',
-    (23, 53, 'U', 'F'): 'U',
-    (23, 53, 'U', 'L'): 'U',
-    (23, 53, 'U', 'R'): 'U',
-    (24, 54, 'B', 'D'): 'D',
-    (24, 54, 'B', 'L'): 'D',
-    (24, 54, 'B', 'R'): 'D',
-    (24, 54, 'B', 'U'): 'D',
-    (24, 54, 'D', 'B'): 'U',
-    (24, 54, 'D', 'F'): 'U',
-    (24, 54, 'D', 'L'): 'U',
-    (24, 54, 'D', 'R'): 'U',
-    (24, 54, 'F', 'D'): 'D',
-    (24, 54, 'F', 'L'): 'D',
-    (24, 54, 'F', 'R'): 'D',
-    (24, 54, 'F', 'U'): 'D',
-    (24, 54, 'L', 'B'): 'U',
-    (24, 54, 'L', 'D'): 'D',
-    (24, 54, 'L', 'F'): 'U',
-    (24, 54, 'L', 'U'): 'D',
-    (24, 54, 'R', 'B'): 'U',
-    (24, 54, 'R', 'D'): 'D',
-    (24, 54, 'R', 'F'): 'U',
-    (24, 54, 'R', 'U'): 'D',
-    (24, 54, 'U', 'B'): 'U',
-    (24, 54, 'U', 'F'): 'U',
-    (24, 54, 'U', 'L'): 'U',
-    (24, 54, 'U', 'R'): 'U',
-    (27, 6, 'B', 'D'): 'D',
-    (27, 6, 'B', 'L'): 'D',
-    (27, 6, 'B', 'R'): 'D',
-    (27, 6, 'B', 'U'): 'D',
-    (27, 6, 'D', 'B'): 'U',
-    (27, 6, 'D', 'F'): 'U',
-    (27, 6, 'D', 'L'): 'U',
-    (27, 6, 'D', 'R'): 'U',
-    (27, 6, 'F', 'D'): 'D',
-    (27, 6, 'F', 'L'): 'D',
-    (27, 6, 'F', 'R'): 'D',
-    (27, 6, 'F', 'U'): 'D',
-    (27, 6, 'L', 'B'): 'U',
-    (27, 6, 'L', 'D'): 'D',
-    (27, 6, 'L', 'F'): 'U',
-    (27, 6, 'L', 'U'): 'D',
-    (27, 6, 'R', 'B'): 'U',
-    (27, 6, 'R', 'D'): 'D',
-    (27, 6, 'R', 'F'): 'U',
-    (27, 6, 'R', 'U'): 'D',
-    (27, 6, 'U', 'B'): 'U',
-    (27, 6, 'U', 'F'): 'U',
-    (27, 6, 'U', 'L'): 'U',
-    (27, 6, 'U', 'R'): 'U',
-    (28, 11, 'B', 'D'): 'U',
-    (28, 11, 'B', 'L'): 'U',
-    (28, 11, 'B', 'R'): 'U',
-    (28, 11, 'B', 'U'): 'U',
-    (28, 11, 'D', 'B'): 'D',
-    (28, 11, 'D', 'F'): 'D',
-    (28, 11, 'D', 'L'): 'D',
-    (28, 11, 'D', 'R'): 'D',
-    (28, 11, 'F', 'D'): 'U',
-    (28, 11, 'F', 'L'): 'U',
-    (28, 11, 'F', 'R'): 'U',
-    (28, 11, 'F', 'U'): 'U',
-    (28, 11, 'L', 'B'): 'D',
-    (28, 11, 'L', 'D'): 'U',
-    (28, 11, 'L', 'F'): 'D',
-    (28, 11, 'L', 'U'): 'U',
-    (28, 11, 'R', 'B'): 'D',
-    (28, 11, 'R', 'D'): 'U',
-    (28, 11, 'R', 'F'): 'D',
-    (28, 11, 'R', 'U'): 'U',
-    (28, 11, 'U', 'B'): 'D',
-    (28, 11, 'U', 'F'): 'D',
-    (28, 11, 'U', 'L'): 'D',
-    (28, 11, 'U', 'R'): 'D',
-    (29, 16, 'B', 'D'): 'U',
-    (29, 16, 'B', 'L'): 'U',
-    (29, 16, 'B', 'R'): 'U',
-    (29, 16, 'B', 'U'): 'U',
-    (29, 16, 'D', 'B'): 'D',
-    (29, 16, 'D', 'F'): 'D',
-    (29, 16, 'D', 'L'): 'D',
-    (29, 16, 'D', 'R'): 'D',
-    (29, 16, 'F', 'D'): 'U',
-    (29, 16, 'F', 'L'): 'U',
-    (29, 16, 'F', 'R'): 'U',
-    (29, 16, 'F', 'U'): 'U',
-    (29, 16, 'L', 'B'): 'D',
-    (29, 16, 'L', 'D'): 'U',
-    (29, 16, 'L', 'F'): 'D',
-    (29, 16, 'L', 'U'): 'U',
-    (29, 16, 'R', 'B'): 'D',
-    (29, 16, 'R', 'D'): 'U',
-    (29, 16, 'R', 'F'): 'D',
-    (29, 16, 'R', 'U'): 'U',
-    (29, 16, 'U', 'B'): 'D',
-    (29, 16, 'U', 'F'): 'D',
-    (29, 16, 'U', 'L'): 'D',
-    (29, 16, 'U', 'R'): 'D',
-    (31, 110, 'B', 'D'): 'U',
-    (31, 110, 'B', 'L'): 'U',
-    (31, 110, 'B', 'R'): 'U',
-    (31, 110, 'B', 'U'): 'U',
-    (31, 110, 'D', 'B'): 'D',
-    (31, 110, 'D', 'F'): 'D',
-    (31, 110, 'D', 'L'): 'D',
-    (31, 110, 'D', 'R'): 'D',
-    (31, 110, 'F', 'D'): 'U',
-    (31, 110, 'F', 'L'): 'U',
-    (31, 110, 'F', 'R'): 'U',
-    (31, 110, 'F', 'U'): 'U',
-    (31, 110, 'L', 'B'): 'D',
-    (31, 110, 'L', 'D'): 'U',
-    (31, 110, 'L', 'F'): 'D',
-    (31, 110, 'L', 'U'): 'U',
-    (31, 110, 'R', 'B'): 'D',
-    (31, 110, 'R', 'D'): 'U',
-    (31, 110, 'R', 'F'): 'D',
-    (31, 110, 'R', 'U'): 'U',
-    (31, 110, 'U', 'B'): 'D',
-    (31, 110, 'U', 'F'): 'D',
-    (31, 110, 'U', 'L'): 'D',
-    (31, 110, 'U', 'R'): 'D',
-    (35, 56, 'B', 'D'): 'D',
-    (35, 56, 'B', 'L'): 'D',
-    (35, 56, 'B', 'R'): 'D',
-    (35, 56, 'B', 'U'): 'D',
-    (35, 56, 'D', 'B'): 'U',
-    (35, 56, 'D', 'F'): 'U',
-    (35, 56, 'D', 'L'): 'U',
-    (35, 56, 'D', 'R'): 'U',
-    (35, 56, 'F', 'D'): 'D',
-    (35, 56, 'F', 'L'): 'D',
-    (35, 56, 'F', 'R'): 'D',
-    (35, 56, 'F', 'U'): 'D',
-    (35, 56, 'L', 'B'): 'U',
-    (35, 56, 'L', 'D'): 'D',
-    (35, 56, 'L', 'F'): 'U',
-    (35, 56, 'L', 'U'): 'D',
-    (35, 56, 'R', 'B'): 'U',
-    (35, 56, 'R', 'D'): 'D',
-    (35, 56, 'R', 'F'): 'U',
-    (35, 56, 'R', 'U'): 'D',
-    (35, 56, 'U', 'B'): 'U',
-    (35, 56, 'U', 'F'): 'U',
-    (35, 56, 'U', 'L'): 'U',
-    (35, 56, 'U', 'R'): 'U',
-    (36, 115, 'B', 'D'): 'D',
-    (36, 115, 'B', 'L'): 'D',
-    (36, 115, 'B', 'R'): 'D',
-    (36, 115, 'B', 'U'): 'D',
-    (36, 115, 'D', 'B'): 'U',
-    (36, 115, 'D', 'F'): 'U',
-    (36, 115, 'D', 'L'): 'U',
-    (36, 115, 'D', 'R'): 'U',
-    (36, 115, 'F', 'D'): 'D',
-    (36, 115, 'F', 'L'): 'D',
-    (36, 115, 'F', 'R'): 'D',
-    (36, 115, 'F', 'U'): 'D',
-    (36, 115, 'L', 'B'): 'U',
-    (36, 115, 'L', 'D'): 'D',
-    (36, 115, 'L', 'F'): 'U',
-    (36, 115, 'L', 'U'): 'D',
-    (36, 115, 'R', 'B'): 'U',
-    (36, 115, 'R', 'D'): 'D',
-    (36, 115, 'R', 'F'): 'U',
-    (36, 115, 'R', 'U'): 'D',
-    (36, 115, 'U', 'B'): 'U',
-    (36, 115, 'U', 'F'): 'U',
-    (36, 115, 'U', 'L'): 'U',
-    (36, 115, 'U', 'R'): 'U',
-    (40, 61, 'B', 'D'): 'D',
-    (40, 61, 'B', 'L'): 'D',
-    (40, 61, 'B', 'R'): 'D',
-    (40, 61, 'B', 'U'): 'D',
-    (40, 61, 'D', 'B'): 'U',
-    (40, 61, 'D', 'F'): 'U',
-    (40, 61, 'D', 'L'): 'U',
-    (40, 61, 'D', 'R'): 'U',
-    (40, 61, 'F', 'D'): 'D',
-    (40, 61, 'F', 'L'): 'D',
-    (40, 61, 'F', 'R'): 'D',
-    (40, 61, 'F', 'U'): 'D',
-    (40, 61, 'L', 'B'): 'U',
-    (40, 61, 'L', 'D'): 'D',
-    (40, 61, 'L', 'F'): 'U',
-    (40, 61, 'L', 'U'): 'D',
-    (40, 61, 'R', 'B'): 'U',
-    (40, 61, 'R', 'D'): 'D',
-    (40, 61, 'R', 'F'): 'U',
-    (40, 61, 'R', 'U'): 'D',
-    (40, 61, 'U', 'B'): 'U',
-    (40, 61, 'U', 'F'): 'U',
-    (40, 61, 'U', 'L'): 'U',
-    (40, 61, 'U', 'R'): 'U',
-    (41, 120, 'B', 'D'): 'D',
-    (41, 120, 'B', 'L'): 'D',
-    (41, 120, 'B', 'R'): 'D',
-    (41, 120, 'B', 'U'): 'D',
-    (41, 120, 'D', 'B'): 'U',
-    (41, 120, 'D', 'F'): 'U',
-    (41, 120, 'D', 'L'): 'U',
-    (41, 120, 'D', 'R'): 'U',
-    (41, 120, 'F', 'D'): 'D',
-    (41, 120, 'F', 'L'): 'D',
-    (41, 120, 'F', 'R'): 'D',
-    (41, 120, 'F', 'U'): 'D',
-    (41, 120, 'L', 'B'): 'U',
-    (41, 120, 'L', 'D'): 'D',
-    (41, 120, 'L', 'F'): 'U',
-    (41, 120, 'L', 'U'): 'D',
-    (41, 120, 'R', 'B'): 'U',
-    (41, 120, 'R', 'D'): 'D',
-    (41, 120, 'R', 'F'): 'U',
-    (41, 120, 'R', 'U'): 'D',
-    (41, 120, 'U', 'B'): 'U',
-    (41, 120, 'U', 'F'): 'U',
-    (41, 120, 'U', 'L'): 'U',
-    (41, 120, 'U', 'R'): 'U',
-    (45, 66, 'B', 'D'): 'U',
-    (45, 66, 'B', 'L'): 'U',
-    (45, 66, 'B', 'R'): 'U',
-    (45, 66, 'B', 'U'): 'U',
-    (45, 66, 'D', 'B'): 'D',
-    (45, 66, 'D', 'F'): 'D',
-    (45, 66, 'D', 'L'): 'D',
-    (45, 66, 'D', 'R'): 'D',
-    (45, 66, 'F', 'D'): 'U',
-    (45, 66, 'F', 'L'): 'U',
-    (45, 66, 'F', 'R'): 'U',
-    (45, 66, 'F', 'U'): 'U',
-    (45, 66, 'L', 'B'): 'D',
-    (45, 66, 'L', 'D'): 'U',
-    (45, 66, 'L', 'F'): 'D',
-    (45, 66, 'L', 'U'): 'U',
-    (45, 66, 'R', 'B'): 'D',
-    (45, 66, 'R', 'D'): 'U',
-    (45, 66, 'R', 'F'): 'D',
-    (45, 66, 'R', 'U'): 'U',
-    (45, 66, 'U', 'B'): 'D',
-    (45, 66, 'U', 'F'): 'D',
-    (45, 66, 'U', 'L'): 'D',
-    (45, 66, 'U', 'R'): 'D',
-    (47, 141, 'B', 'D'): 'U',
-    (47, 141, 'B', 'L'): 'U',
-    (47, 141, 'B', 'R'): 'U',
-    (47, 141, 'B', 'U'): 'U',
-    (47, 141, 'D', 'B'): 'D',
-    (47, 141, 'D', 'F'): 'D',
-    (47, 141, 'D', 'L'): 'D',
-    (47, 141, 'D', 'R'): 'D',
-    (47, 141, 'F', 'D'): 'U',
-    (47, 141, 'F', 'L'): 'U',
-    (47, 141, 'F', 'R'): 'U',
-    (47, 141, 'F', 'U'): 'U',
-    (47, 141, 'L', 'B'): 'D',
-    (47, 141, 'L', 'D'): 'U',
-    (47, 141, 'L', 'F'): 'D',
-    (47, 141, 'L', 'U'): 'U',
-    (47, 141, 'R', 'B'): 'D',
-    (47, 141, 'R', 'D'): 'U',
-    (47, 141, 'R', 'F'): 'D',
-    (47, 141, 'R', 'U'): 'U',
-    (47, 141, 'U', 'B'): 'D',
-    (47, 141, 'U', 'F'): 'D',
-    (47, 141, 'U', 'L'): 'D',
-    (47, 141, 'U', 'R'): 'D',
-    (48, 136, 'B', 'D'): 'U',
-    (48, 136, 'B', 'L'): 'U',
-    (48, 136, 'B', 'R'): 'U',
-    (48, 136, 'B', 'U'): 'U',
-    (48, 136, 'D', 'B'): 'D',
-    (48, 136, 'D', 'F'): 'D',
-    (48, 136, 'D', 'L'): 'D',
-    (48, 136, 'D', 'R'): 'D',
-    (48, 136, 'F', 'D'): 'U',
-    (48, 136, 'F', 'L'): 'U',
-    (48, 136, 'F', 'R'): 'U',
-    (48, 136, 'F', 'U'): 'U',
-    (48, 136, 'L', 'B'): 'D',
-    (48, 136, 'L', 'D'): 'U',
-    (48, 136, 'L', 'F'): 'D',
-    (48, 136, 'L', 'U'): 'U',
-    (48, 136, 'R', 'B'): 'D',
-    (48, 136, 'R', 'D'): 'U',
-    (48, 136, 'R', 'F'): 'D',
-    (48, 136, 'R', 'U'): 'U',
-    (48, 136, 'U', 'B'): 'D',
-    (48, 136, 'U', 'F'): 'D',
-    (48, 136, 'U', 'L'): 'D',
-    (48, 136, 'U', 'R'): 'D',
-    (49, 131, 'B', 'D'): 'D',
-    (49, 131, 'B', 'L'): 'D',
-    (49, 131, 'B', 'R'): 'D',
-    (49, 131, 'B', 'U'): 'D',
-    (49, 131, 'D', 'B'): 'U',
-    (49, 131, 'D', 'F'): 'U',
-    (49, 131, 'D', 'L'): 'U',
-    (49, 131, 'D', 'R'): 'U',
-    (49, 131, 'F', 'D'): 'D',
-    (49, 131, 'F', 'L'): 'D',
-    (49, 131, 'F', 'R'): 'D',
-    (49, 131, 'F', 'U'): 'D',
-    (49, 131, 'L', 'B'): 'U',
-    (49, 131, 'L', 'D'): 'D',
-    (49, 131, 'L', 'F'): 'U',
-    (49, 131, 'L', 'U'): 'D',
-    (49, 131, 'R', 'B'): 'U',
-    (49, 131, 'R', 'D'): 'D',
-    (49, 131, 'R', 'F'): 'U',
-    (49, 131, 'R', 'U'): 'D',
-    (49, 131, 'U', 'B'): 'U',
-    (49, 131, 'U', 'F'): 'U',
-    (49, 131, 'U', 'L'): 'U',
-    (49, 131, 'U', 'R'): 'U',
-    (52, 22, 'B', 'D'): 'D',
-    (52, 22, 'B', 'L'): 'D',
-    (52, 22, 'B', 'R'): 'D',
-    (52, 22, 'B', 'U'): 'D',
-    (52, 22, 'D', 'B'): 'U',
-    (52, 22, 'D', 'F'): 'U',
-    (52, 22, 'D', 'L'): 'U',
-    (52, 22, 'D', 'R'): 'U',
-    (52, 22, 'F', 'D'): 'D',
-    (52, 22, 'F', 'L'): 'D',
-    (52, 22, 'F', 'R'): 'D',
-    (52, 22, 'F', 'U'): 'D',
-    (52, 22, 'L', 'B'): 'U',
-    (52, 22, 'L', 'D'): 'D',
-    (52, 22, 'L', 'F'): 'U',
-    (52, 22, 'L', 'U'): 'D',
-    (52, 22, 'R', 'B'): 'U',
-    (52, 22, 'R', 'D'): 'D',
-    (52, 22, 'R', 'F'): 'U',
-    (52, 22, 'R', 'U'): 'D',
-    (52, 22, 'U', 'B'): 'U',
-    (52, 22, 'U', 'F'): 'U',
-    (52, 22, 'U', 'L'): 'U',
-    (52, 22, 'U', 'R'): 'U',
-    (53, 23, 'B', 'D'): 'U',
-    (53, 23, 'B', 'L'): 'U',
-    (53, 23, 'B', 'R'): 'U',
-    (53, 23, 'B', 'U'): 'U',
-    (53, 23, 'D', 'B'): 'D',
-    (53, 23, 'D', 'F'): 'D',
-    (53, 23, 'D', 'L'): 'D',
-    (53, 23, 'D', 'R'): 'D',
-    (53, 23, 'F', 'D'): 'U',
-    (53, 23, 'F', 'L'): 'U',
-    (53, 23, 'F', 'R'): 'U',
-    (53, 23, 'F', 'U'): 'U',
-    (53, 23, 'L', 'B'): 'D',
-    (53, 23, 'L', 'D'): 'U',
-    (53, 23, 'L', 'F'): 'D',
-    (53, 23, 'L', 'U'): 'U',
-    (53, 23, 'R', 'B'): 'D',
-    (53, 23, 'R', 'D'): 'U',
-    (53, 23, 'R', 'F'): 'D',
-    (53, 23, 'R', 'U'): 'U',
-    (53, 23, 'U', 'B'): 'D',
-    (53, 23, 'U', 'F'): 'D',
-    (53, 23, 'U', 'L'): 'D',
-    (53, 23, 'U', 'R'): 'D',
-    (54, 24, 'B', 'D'): 'U',
-    (54, 24, 'B', 'L'): 'U',
-    (54, 24, 'B', 'R'): 'U',
-    (54, 24, 'B', 'U'): 'U',
-    (54, 24, 'D', 'B'): 'D',
-    (54, 24, 'D', 'F'): 'D',
-    (54, 24, 'D', 'L'): 'D',
-    (54, 24, 'D', 'R'): 'D',
-    (54, 24, 'F', 'D'): 'U',
-    (54, 24, 'F', 'L'): 'U',
-    (54, 24, 'F', 'R'): 'U',
-    (54, 24, 'F', 'U'): 'U',
-    (54, 24, 'L', 'B'): 'D',
-    (54, 24, 'L', 'D'): 'U',
-    (54, 24, 'L', 'F'): 'D',
-    (54, 24, 'L', 'U'): 'U',
-    (54, 24, 'R', 'B'): 'D',
-    (54, 24, 'R', 'D'): 'U',
-    (54, 24, 'R', 'F'): 'D',
-    (54, 24, 'R', 'U'): 'U',
-    (54, 24, 'U', 'B'): 'D',
-    (54, 24, 'U', 'F'): 'D',
-    (54, 24, 'U', 'L'): 'D',
-    (54, 24, 'U', 'R'): 'D',
-    (56, 35, 'B', 'D'): 'U',
-    (56, 35, 'B', 'L'): 'U',
-    (56, 35, 'B', 'R'): 'U',
-    (56, 35, 'B', 'U'): 'U',
-    (56, 35, 'D', 'B'): 'D',
-    (56, 35, 'D', 'F'): 'D',
-    (56, 35, 'D', 'L'): 'D',
-    (56, 35, 'D', 'R'): 'D',
-    (56, 35, 'F', 'D'): 'U',
-    (56, 35, 'F', 'L'): 'U',
-    (56, 35, 'F', 'R'): 'U',
-    (56, 35, 'F', 'U'): 'U',
-    (56, 35, 'L', 'B'): 'D',
-    (56, 35, 'L', 'D'): 'U',
-    (56, 35, 'L', 'F'): 'D',
-    (56, 35, 'L', 'U'): 'U',
-    (56, 35, 'R', 'B'): 'D',
-    (56, 35, 'R', 'D'): 'U',
-    (56, 35, 'R', 'F'): 'D',
-    (56, 35, 'R', 'U'): 'U',
-    (56, 35, 'U', 'B'): 'D',
-    (56, 35, 'U', 'F'): 'D',
-    (56, 35, 'U', 'L'): 'D',
-    (56, 35, 'U', 'R'): 'D',
-    (60, 81, 'B', 'D'): 'D',
-    (60, 81, 'B', 'L'): 'D',
-    (60, 81, 'B', 'R'): 'D',
-    (60, 81, 'B', 'U'): 'D',
-    (60, 81, 'D', 'B'): 'U',
-    (60, 81, 'D', 'F'): 'U',
-    (60, 81, 'D', 'L'): 'U',
-    (60, 81, 'D', 'R'): 'U',
-    (60, 81, 'F', 'D'): 'D',
-    (60, 81, 'F', 'L'): 'D',
-    (60, 81, 'F', 'R'): 'D',
-    (60, 81, 'F', 'U'): 'D',
-    (60, 81, 'L', 'B'): 'U',
-    (60, 81, 'L', 'D'): 'D',
-    (60, 81, 'L', 'F'): 'U',
-    (60, 81, 'L', 'U'): 'D',
-    (60, 81, 'R', 'B'): 'U',
-    (60, 81, 'R', 'D'): 'D',
-    (60, 81, 'R', 'F'): 'U',
-    (60, 81, 'R', 'U'): 'D',
-    (60, 81, 'U', 'B'): 'U',
-    (60, 81, 'U', 'F'): 'U',
-    (60, 81, 'U', 'L'): 'U',
-    (60, 81, 'U', 'R'): 'U',
-    (61, 40, 'B', 'D'): 'U',
-    (61, 40, 'B', 'L'): 'U',
-    (61, 40, 'B', 'R'): 'U',
-    (61, 40, 'B', 'U'): 'U',
-    (61, 40, 'D', 'B'): 'D',
-    (61, 40, 'D', 'F'): 'D',
-    (61, 40, 'D', 'L'): 'D',
-    (61, 40, 'D', 'R'): 'D',
-    (61, 40, 'F', 'D'): 'U',
-    (61, 40, 'F', 'L'): 'U',
-    (61, 40, 'F', 'R'): 'U',
-    (61, 40, 'F', 'U'): 'U',
-    (61, 40, 'L', 'B'): 'D',
-    (61, 40, 'L', 'D'): 'U',
-    (61, 40, 'L', 'F'): 'D',
-    (61, 40, 'L', 'U'): 'U',
-    (61, 40, 'R', 'B'): 'D',
-    (61, 40, 'R', 'D'): 'U',
-    (61, 40, 'R', 'F'): 'D',
-    (61, 40, 'R', 'U'): 'U',
-    (61, 40, 'U', 'B'): 'D',
-    (61, 40, 'U', 'F'): 'D',
-    (61, 40, 'U', 'L'): 'D',
-    (61, 40, 'U', 'R'): 'D',
-    (65, 86, 'B', 'D'): 'U',
-    (65, 86, 'B', 'L'): 'U',
-    (65, 86, 'B', 'R'): 'U',
-    (65, 86, 'B', 'U'): 'U',
-    (65, 86, 'D', 'B'): 'D',
-    (65, 86, 'D', 'F'): 'D',
-    (65, 86, 'D', 'L'): 'D',
-    (65, 86, 'D', 'R'): 'D',
-    (65, 86, 'F', 'D'): 'U',
-    (65, 86, 'F', 'L'): 'U',
-    (65, 86, 'F', 'R'): 'U',
-    (65, 86, 'F', 'U'): 'U',
-    (65, 86, 'L', 'B'): 'D',
-    (65, 86, 'L', 'D'): 'U',
-    (65, 86, 'L', 'F'): 'D',
-    (65, 86, 'L', 'U'): 'U',
-    (65, 86, 'R', 'B'): 'D',
-    (65, 86, 'R', 'D'): 'U',
-    (65, 86, 'R', 'F'): 'D',
-    (65, 86, 'R', 'U'): 'U',
-    (65, 86, 'U', 'B'): 'D',
-    (65, 86, 'U', 'F'): 'D',
-    (65, 86, 'U', 'L'): 'D',
-    (65, 86, 'U', 'R'): 'D',
-    (66, 45, 'B', 'D'): 'D',
-    (66, 45, 'B', 'L'): 'D',
-    (66, 45, 'B', 'R'): 'D',
-    (66, 45, 'B', 'U'): 'D',
-    (66, 45, 'D', 'B'): 'U',
-    (66, 45, 'D', 'F'): 'U',
-    (66, 45, 'D', 'L'): 'U',
-    (66, 45, 'D', 'R'): 'U',
-    (66, 45, 'F', 'D'): 'D',
-    (66, 45, 'F', 'L'): 'D',
-    (66, 45, 'F', 'R'): 'D',
-    (66, 45, 'F', 'U'): 'D',
-    (66, 45, 'L', 'B'): 'U',
-    (66, 45, 'L', 'D'): 'D',
-    (66, 45, 'L', 'F'): 'U',
-    (66, 45, 'L', 'U'): 'D',
-    (66, 45, 'R', 'B'): 'U',
-    (66, 45, 'R', 'D'): 'D',
-    (66, 45, 'R', 'F'): 'U',
-    (66, 45, 'R', 'U'): 'D',
-    (66, 45, 'U', 'B'): 'U',
-    (66, 45, 'U', 'F'): 'U',
-    (66, 45, 'U', 'L'): 'U',
-    (66, 45, 'U', 'R'): 'U',
-    (70, 91, 'B', 'D'): 'U',
-    (70, 91, 'B', 'L'): 'U',
-    (70, 91, 'B', 'R'): 'U',
-    (70, 91, 'B', 'U'): 'U',
-    (70, 91, 'D', 'B'): 'D',
-    (70, 91, 'D', 'F'): 'D',
-    (70, 91, 'D', 'L'): 'D',
-    (70, 91, 'D', 'R'): 'D',
-    (70, 91, 'F', 'D'): 'U',
-    (70, 91, 'F', 'L'): 'U',
-    (70, 91, 'F', 'R'): 'U',
-    (70, 91, 'F', 'U'): 'U',
-    (70, 91, 'L', 'B'): 'D',
-    (70, 91, 'L', 'D'): 'U',
-    (70, 91, 'L', 'F'): 'D',
-    (70, 91, 'L', 'U'): 'U',
-    (70, 91, 'R', 'B'): 'D',
-    (70, 91, 'R', 'D'): 'U',
-    (70, 91, 'R', 'F'): 'D',
-    (70, 91, 'R', 'U'): 'U',
-    (70, 91, 'U', 'B'): 'D',
-    (70, 91, 'U', 'F'): 'D',
-    (70, 91, 'U', 'L'): 'D',
-    (70, 91, 'U', 'R'): 'D',
-    (72, 127, 'B', 'D'): 'U',
-    (72, 127, 'B', 'L'): 'U',
-    (72, 127, 'B', 'R'): 'U',
-    (72, 127, 'B', 'U'): 'U',
-    (72, 127, 'D', 'B'): 'D',
-    (72, 127, 'D', 'F'): 'D',
-    (72, 127, 'D', 'L'): 'D',
-    (72, 127, 'D', 'R'): 'D',
-    (72, 127, 'F', 'D'): 'U',
-    (72, 127, 'F', 'L'): 'U',
-    (72, 127, 'F', 'R'): 'U',
-    (72, 127, 'F', 'U'): 'U',
-    (72, 127, 'L', 'B'): 'D',
-    (72, 127, 'L', 'D'): 'U',
-    (72, 127, 'L', 'F'): 'D',
-    (72, 127, 'L', 'U'): 'U',
-    (72, 127, 'R', 'B'): 'D',
-    (72, 127, 'R', 'D'): 'U',
-    (72, 127, 'R', 'F'): 'D',
-    (72, 127, 'R', 'U'): 'U',
-    (72, 127, 'U', 'B'): 'D',
-    (72, 127, 'U', 'F'): 'D',
-    (72, 127, 'U', 'L'): 'D',
-    (72, 127, 'U', 'R'): 'D',
-    (73, 128, 'B', 'D'): 'U',
-    (73, 128, 'B', 'L'): 'U',
-    (73, 128, 'B', 'R'): 'U',
-    (73, 128, 'B', 'U'): 'U',
-    (73, 128, 'D', 'B'): 'D',
-    (73, 128, 'D', 'F'): 'D',
-    (73, 128, 'D', 'L'): 'D',
-    (73, 128, 'D', 'R'): 'D',
-    (73, 128, 'F', 'D'): 'U',
-    (73, 128, 'F', 'L'): 'U',
-    (73, 128, 'F', 'R'): 'U',
-    (73, 128, 'F', 'U'): 'U',
-    (73, 128, 'L', 'B'): 'D',
-    (73, 128, 'L', 'D'): 'U',
-    (73, 128, 'L', 'F'): 'D',
-    (73, 128, 'L', 'U'): 'U',
-    (73, 128, 'R', 'B'): 'D',
-    (73, 128, 'R', 'D'): 'U',
-    (73, 128, 'R', 'F'): 'D',
-    (73, 128, 'R', 'U'): 'U',
-    (73, 128, 'U', 'B'): 'D',
-    (73, 128, 'U', 'F'): 'D',
-    (73, 128, 'U', 'L'): 'D',
-    (73, 128, 'U', 'R'): 'D',
-    (74, 129, 'B', 'D'): 'D',
-    (74, 129, 'B', 'L'): 'D',
-    (74, 129, 'B', 'R'): 'D',
-    (74, 129, 'B', 'U'): 'D',
-    (74, 129, 'D', 'B'): 'U',
-    (74, 129, 'D', 'F'): 'U',
-    (74, 129, 'D', 'L'): 'U',
-    (74, 129, 'D', 'R'): 'U',
-    (74, 129, 'F', 'D'): 'D',
-    (74, 129, 'F', 'L'): 'D',
-    (74, 129, 'F', 'R'): 'D',
-    (74, 129, 'F', 'U'): 'D',
-    (74, 129, 'L', 'B'): 'U',
-    (74, 129, 'L', 'D'): 'D',
-    (74, 129, 'L', 'F'): 'U',
-    (74, 129, 'L', 'U'): 'D',
-    (74, 129, 'R', 'B'): 'U',
-    (74, 129, 'R', 'D'): 'D',
-    (74, 129, 'R', 'F'): 'U',
-    (74, 129, 'R', 'U'): 'D',
-    (74, 129, 'U', 'B'): 'U',
-    (74, 129, 'U', 'F'): 'U',
-    (74, 129, 'U', 'L'): 'U',
-    (74, 129, 'U', 'R'): 'U',
-    (77, 20, 'B', 'D'): 'D',
-    (77, 20, 'B', 'L'): 'D',
-    (77, 20, 'B', 'R'): 'D',
-    (77, 20, 'B', 'U'): 'D',
-    (77, 20, 'D', 'B'): 'U',
-    (77, 20, 'D', 'F'): 'U',
-    (77, 20, 'D', 'L'): 'U',
-    (77, 20, 'D', 'R'): 'U',
-    (77, 20, 'F', 'D'): 'D',
-    (77, 20, 'F', 'L'): 'D',
-    (77, 20, 'F', 'R'): 'D',
-    (77, 20, 'F', 'U'): 'D',
-    (77, 20, 'L', 'B'): 'U',
-    (77, 20, 'L', 'D'): 'D',
-    (77, 20, 'L', 'F'): 'U',
-    (77, 20, 'L', 'U'): 'D',
-    (77, 20, 'R', 'B'): 'U',
-    (77, 20, 'R', 'D'): 'D',
-    (77, 20, 'R', 'F'): 'U',
-    (77, 20, 'R', 'U'): 'D',
-    (77, 20, 'U', 'B'): 'U',
-    (77, 20, 'U', 'F'): 'U',
-    (77, 20, 'U', 'L'): 'U',
-    (77, 20, 'U', 'R'): 'U',
-    (78, 15, 'B', 'D'): 'U',
-    (78, 15, 'B', 'L'): 'U',
-    (78, 15, 'B', 'R'): 'U',
-    (78, 15, 'B', 'U'): 'U',
-    (78, 15, 'D', 'B'): 'D',
-    (78, 15, 'D', 'F'): 'D',
-    (78, 15, 'D', 'L'): 'D',
-    (78, 15, 'D', 'R'): 'D',
-    (78, 15, 'F', 'D'): 'U',
-    (78, 15, 'F', 'L'): 'U',
-    (78, 15, 'F', 'R'): 'U',
-    (78, 15, 'F', 'U'): 'U',
-    (78, 15, 'L', 'B'): 'D',
-    (78, 15, 'L', 'D'): 'U',
-    (78, 15, 'L', 'F'): 'D',
-    (78, 15, 'L', 'U'): 'U',
-    (78, 15, 'R', 'B'): 'D',
-    (78, 15, 'R', 'D'): 'U',
-    (78, 15, 'R', 'F'): 'D',
-    (78, 15, 'R', 'U'): 'U',
-    (78, 15, 'U', 'B'): 'D',
-    (78, 15, 'U', 'F'): 'D',
-    (78, 15, 'U', 'L'): 'D',
-    (78, 15, 'U', 'R'): 'D',
-    (79, 10, 'B', 'D'): 'U',
-    (79, 10, 'B', 'L'): 'U',
-    (79, 10, 'B', 'R'): 'U',
-    (79, 10, 'B', 'U'): 'U',
-    (79, 10, 'D', 'B'): 'D',
-    (79, 10, 'D', 'F'): 'D',
-    (79, 10, 'D', 'L'): 'D',
-    (79, 10, 'D', 'R'): 'D',
-    (79, 10, 'F', 'D'): 'U',
-    (79, 10, 'F', 'L'): 'U',
-    (79, 10, 'F', 'R'): 'U',
-    (79, 10, 'F', 'U'): 'U',
-    (79, 10, 'L', 'B'): 'D',
-    (79, 10, 'L', 'D'): 'U',
-    (79, 10, 'L', 'F'): 'D',
-    (79, 10, 'L', 'U'): 'U',
-    (79, 10, 'R', 'B'): 'D',
-    (79, 10, 'R', 'D'): 'U',
-    (79, 10, 'R', 'F'): 'D',
-    (79, 10, 'R', 'U'): 'U',
-    (79, 10, 'U', 'B'): 'D',
-    (79, 10, 'U', 'F'): 'D',
-    (79, 10, 'U', 'L'): 'D',
-    (79, 10, 'U', 'R'): 'D',
-    (81, 60, 'B', 'D'): 'U',
-    (81, 60, 'B', 'L'): 'U',
-    (81, 60, 'B', 'R'): 'U',
-    (81, 60, 'B', 'U'): 'U',
-    (81, 60, 'D', 'B'): 'D',
-    (81, 60, 'D', 'F'): 'D',
-    (81, 60, 'D', 'L'): 'D',
-    (81, 60, 'D', 'R'): 'D',
-    (81, 60, 'F', 'D'): 'U',
-    (81, 60, 'F', 'L'): 'U',
-    (81, 60, 'F', 'R'): 'U',
-    (81, 60, 'F', 'U'): 'U',
-    (81, 60, 'L', 'B'): 'D',
-    (81, 60, 'L', 'D'): 'U',
-    (81, 60, 'L', 'F'): 'D',
-    (81, 60, 'L', 'U'): 'U',
-    (81, 60, 'R', 'B'): 'D',
-    (81, 60, 'R', 'D'): 'U',
-    (81, 60, 'R', 'F'): 'D',
-    (81, 60, 'R', 'U'): 'U',
-    (81, 60, 'U', 'B'): 'D',
-    (81, 60, 'U', 'F'): 'D',
-    (81, 60, 'U', 'L'): 'D',
-    (81, 60, 'U', 'R'): 'D',
-    (85, 106, 'B', 'D'): 'D',
-    (85, 106, 'B', 'L'): 'D',
-    (85, 106, 'B', 'R'): 'D',
-    (85, 106, 'B', 'U'): 'D',
-    (85, 106, 'D', 'B'): 'U',
-    (85, 106, 'D', 'F'): 'U',
-    (85, 106, 'D', 'L'): 'U',
-    (85, 106, 'D', 'R'): 'U',
-    (85, 106, 'F', 'D'): 'D',
-    (85, 106, 'F', 'L'): 'D',
-    (85, 106, 'F', 'R'): 'D',
-    (85, 106, 'F', 'U'): 'D',
-    (85, 106, 'L', 'B'): 'U',
-    (85, 106, 'L', 'D'): 'D',
-    (85, 106, 'L', 'F'): 'U',
-    (85, 106, 'L', 'U'): 'D',
-    (85, 106, 'R', 'B'): 'U',
-    (85, 106, 'R', 'D'): 'D',
-    (85, 106, 'R', 'F'): 'U',
-    (85, 106, 'R', 'U'): 'D',
-    (85, 106, 'U', 'B'): 'U',
-    (85, 106, 'U', 'F'): 'U',
-    (85, 106, 'U', 'L'): 'U',
-    (85, 106, 'U', 'R'): 'U',
-    (86, 65, 'B', 'D'): 'D',
-    (86, 65, 'B', 'L'): 'D',
-    (86, 65, 'B', 'R'): 'D',
-    (86, 65, 'B', 'U'): 'D',
-    (86, 65, 'D', 'B'): 'U',
-    (86, 65, 'D', 'F'): 'U',
-    (86, 65, 'D', 'L'): 'U',
-    (86, 65, 'D', 'R'): 'U',
-    (86, 65, 'F', 'D'): 'D',
-    (86, 65, 'F', 'L'): 'D',
-    (86, 65, 'F', 'R'): 'D',
-    (86, 65, 'F', 'U'): 'D',
-    (86, 65, 'L', 'B'): 'U',
-    (86, 65, 'L', 'D'): 'D',
-    (86, 65, 'L', 'F'): 'U',
-    (86, 65, 'L', 'U'): 'D',
-    (86, 65, 'R', 'B'): 'U',
-    (86, 65, 'R', 'D'): 'D',
-    (86, 65, 'R', 'F'): 'U',
-    (86, 65, 'R', 'U'): 'D',
-    (86, 65, 'U', 'B'): 'U',
-    (86, 65, 'U', 'F'): 'U',
-    (86, 65, 'U', 'L'): 'U',
-    (86, 65, 'U', 'R'): 'U',
-    (90, 111, 'B', 'D'): 'D',
-    (90, 111, 'B', 'L'): 'D',
-    (90, 111, 'B', 'R'): 'D',
-    (90, 111, 'B', 'U'): 'D',
-    (90, 111, 'D', 'B'): 'U',
-    (90, 111, 'D', 'F'): 'U',
-    (90, 111, 'D', 'L'): 'U',
-    (90, 111, 'D', 'R'): 'U',
-    (90, 111, 'F', 'D'): 'D',
-    (90, 111, 'F', 'L'): 'D',
-    (90, 111, 'F', 'R'): 'D',
-    (90, 111, 'F', 'U'): 'D',
-    (90, 111, 'L', 'B'): 'U',
-    (90, 111, 'L', 'D'): 'D',
-    (90, 111, 'L', 'F'): 'U',
-    (90, 111, 'L', 'U'): 'D',
-    (90, 111, 'R', 'B'): 'U',
-    (90, 111, 'R', 'D'): 'D',
-    (90, 111, 'R', 'F'): 'U',
-    (90, 111, 'R', 'U'): 'D',
-    (90, 111, 'U', 'B'): 'U',
-    (90, 111, 'U', 'F'): 'U',
-    (90, 111, 'U', 'L'): 'U',
-    (90, 111, 'U', 'R'): 'U',
-    (91, 70, 'B', 'D'): 'D',
-    (91, 70, 'B', 'L'): 'D',
-    (91, 70, 'B', 'R'): 'D',
-    (91, 70, 'B', 'U'): 'D',
-    (91, 70, 'D', 'B'): 'U',
-    (91, 70, 'D', 'F'): 'U',
-    (91, 70, 'D', 'L'): 'U',
-    (91, 70, 'D', 'R'): 'U',
-    (91, 70, 'F', 'D'): 'D',
-    (91, 70, 'F', 'L'): 'D',
-    (91, 70, 'F', 'R'): 'D',
-    (91, 70, 'F', 'U'): 'D',
-    (91, 70, 'L', 'B'): 'U',
-    (91, 70, 'L', 'D'): 'D',
-    (91, 70, 'L', 'F'): 'U',
-    (91, 70, 'L', 'U'): 'D',
-    (91, 70, 'R', 'B'): 'U',
-    (91, 70, 'R', 'D'): 'D',
-    (91, 70, 'R', 'F'): 'U',
-    (91, 70, 'R', 'U'): 'D',
-    (91, 70, 'U', 'B'): 'U',
-    (91, 70, 'U', 'F'): 'U',
-    (91, 70, 'U', 'L'): 'U',
-    (91, 70, 'U', 'R'): 'U',
-    (95, 116, 'B', 'D'): 'U',
-    (95, 116, 'B', 'L'): 'U',
-    (95, 116, 'B', 'R'): 'U',
-    (95, 116, 'B', 'U'): 'U',
-    (95, 116, 'D', 'B'): 'D',
-    (95, 116, 'D', 'F'): 'D',
-    (95, 116, 'D', 'L'): 'D',
-    (95, 116, 'D', 'R'): 'D',
-    (95, 116, 'F', 'D'): 'U',
-    (95, 116, 'F', 'L'): 'U',
-    (95, 116, 'F', 'R'): 'U',
-    (95, 116, 'F', 'U'): 'U',
-    (95, 116, 'L', 'B'): 'D',
-    (95, 116, 'L', 'D'): 'U',
-    (95, 116, 'L', 'F'): 'D',
-    (95, 116, 'L', 'U'): 'U',
-    (95, 116, 'R', 'B'): 'D',
-    (95, 116, 'R', 'D'): 'U',
-    (95, 116, 'R', 'F'): 'D',
-    (95, 116, 'R', 'U'): 'U',
-    (95, 116, 'U', 'B'): 'D',
-    (95, 116, 'U', 'F'): 'D',
-    (95, 116, 'U', 'L'): 'D',
-    (95, 116, 'U', 'R'): 'D',
-    (97, 135, 'B', 'D'): 'U',
-    (97, 135, 'B', 'L'): 'U',
-    (97, 135, 'B', 'R'): 'U',
-    (97, 135, 'B', 'U'): 'U',
-    (97, 135, 'D', 'B'): 'D',
-    (97, 135, 'D', 'F'): 'D',
-    (97, 135, 'D', 'L'): 'D',
-    (97, 135, 'D', 'R'): 'D',
-    (97, 135, 'F', 'D'): 'U',
-    (97, 135, 'F', 'L'): 'U',
-    (97, 135, 'F', 'R'): 'U',
-    (97, 135, 'F', 'U'): 'U',
-    (97, 135, 'L', 'B'): 'D',
-    (97, 135, 'L', 'D'): 'U',
-    (97, 135, 'L', 'F'): 'D',
-    (97, 135, 'L', 'U'): 'U',
-    (97, 135, 'R', 'B'): 'D',
-    (97, 135, 'R', 'D'): 'U',
-    (97, 135, 'R', 'F'): 'D',
-    (97, 135, 'R', 'U'): 'U',
-    (97, 135, 'U', 'B'): 'D',
-    (97, 135, 'U', 'F'): 'D',
-    (97, 135, 'U', 'L'): 'D',
-    (97, 135, 'U', 'R'): 'D',
-    (98, 140, 'B', 'D'): 'U',
-    (98, 140, 'B', 'L'): 'U',
-    (98, 140, 'B', 'R'): 'U',
-    (98, 140, 'B', 'U'): 'U',
-    (98, 140, 'D', 'B'): 'D',
-    (98, 140, 'D', 'F'): 'D',
-    (98, 140, 'D', 'L'): 'D',
-    (98, 140, 'D', 'R'): 'D',
-    (98, 140, 'F', 'D'): 'U',
-    (98, 140, 'F', 'L'): 'U',
-    (98, 140, 'F', 'R'): 'U',
-    (98, 140, 'F', 'U'): 'U',
-    (98, 140, 'L', 'B'): 'D',
-    (98, 140, 'L', 'D'): 'U',
-    (98, 140, 'L', 'F'): 'D',
-    (98, 140, 'L', 'U'): 'U',
-    (98, 140, 'R', 'B'): 'D',
-    (98, 140, 'R', 'D'): 'U',
-    (98, 140, 'R', 'F'): 'D',
-    (98, 140, 'R', 'U'): 'U',
-    (98, 140, 'U', 'B'): 'D',
-    (98, 140, 'U', 'F'): 'D',
-    (98, 140, 'U', 'L'): 'D',
-    (98, 140, 'U', 'R'): 'D',
-    (99, 145, 'B', 'D'): 'D',
-    (99, 145, 'B', 'L'): 'D',
-    (99, 145, 'B', 'R'): 'D',
-    (99, 145, 'B', 'U'): 'D',
-    (99, 145, 'D', 'B'): 'U',
-    (99, 145, 'D', 'F'): 'U',
-    (99, 145, 'D', 'L'): 'U',
-    (99, 145, 'D', 'R'): 'U',
-    (99, 145, 'F', 'D'): 'D',
-    (99, 145, 'F', 'L'): 'D',
-    (99, 145, 'F', 'R'): 'D',
-    (99, 145, 'F', 'U'): 'D',
-    (99, 145, 'L', 'B'): 'U',
-    (99, 145, 'L', 'D'): 'D',
-    (99, 145, 'L', 'F'): 'U',
-    (99, 145, 'L', 'U'): 'D',
-    (99, 145, 'R', 'B'): 'U',
-    (99, 145, 'R', 'D'): 'D',
-    (99, 145, 'R', 'F'): 'U',
-    (99, 145, 'R', 'U'): 'D',
-    (99, 145, 'U', 'B'): 'U',
-    (99, 145, 'U', 'F'): 'U',
-    (99, 145, 'U', 'L'): 'U',
-    (99, 145, 'U', 'R'): 'U',
-    (102, 4, 'B', 'D'): 'D',
-    (102, 4, 'B', 'L'): 'D',
-    (102, 4, 'B', 'R'): 'D',
-    (102, 4, 'B', 'U'): 'D',
-    (102, 4, 'D', 'B'): 'U',
-    (102, 4, 'D', 'F'): 'U',
-    (102, 4, 'D', 'L'): 'U',
-    (102, 4, 'D', 'R'): 'U',
-    (102, 4, 'F', 'D'): 'D',
-    (102, 4, 'F', 'L'): 'D',
-    (102, 4, 'F', 'R'): 'D',
-    (102, 4, 'F', 'U'): 'D',
-    (102, 4, 'L', 'B'): 'U',
-    (102, 4, 'L', 'D'): 'D',
-    (102, 4, 'L', 'F'): 'U',
-    (102, 4, 'L', 'U'): 'D',
-    (102, 4, 'R', 'B'): 'U',
-    (102, 4, 'R', 'D'): 'D',
-    (102, 4, 'R', 'F'): 'U',
-    (102, 4, 'R', 'U'): 'D',
-    (102, 4, 'U', 'B'): 'U',
-    (102, 4, 'U', 'F'): 'U',
-    (102, 4, 'U', 'L'): 'U',
-    (102, 4, 'U', 'R'): 'U',
-    (103, 3, 'B', 'D'): 'U',
-    (103, 3, 'B', 'L'): 'U',
-    (103, 3, 'B', 'R'): 'U',
-    (103, 3, 'B', 'U'): 'U',
-    (103, 3, 'D', 'B'): 'D',
-    (103, 3, 'D', 'F'): 'D',
-    (103, 3, 'D', 'L'): 'D',
-    (103, 3, 'D', 'R'): 'D',
-    (103, 3, 'F', 'D'): 'U',
-    (103, 3, 'F', 'L'): 'U',
-    (103, 3, 'F', 'R'): 'U',
-    (103, 3, 'F', 'U'): 'U',
-    (103, 3, 'L', 'B'): 'D',
-    (103, 3, 'L', 'D'): 'U',
-    (103, 3, 'L', 'F'): 'D',
-    (103, 3, 'L', 'U'): 'U',
-    (103, 3, 'R', 'B'): 'D',
-    (103, 3, 'R', 'D'): 'U',
-    (103, 3, 'R', 'F'): 'D',
-    (103, 3, 'R', 'U'): 'U',
-    (103, 3, 'U', 'B'): 'D',
-    (103, 3, 'U', 'F'): 'D',
-    (103, 3, 'U', 'L'): 'D',
-    (103, 3, 'U', 'R'): 'D',
-    (104, 2, 'B', 'D'): 'U',
-    (104, 2, 'B', 'L'): 'U',
-    (104, 2, 'B', 'R'): 'U',
-    (104, 2, 'B', 'U'): 'U',
-    (104, 2, 'D', 'B'): 'D',
-    (104, 2, 'D', 'F'): 'D',
-    (104, 2, 'D', 'L'): 'D',
-    (104, 2, 'D', 'R'): 'D',
-    (104, 2, 'F', 'D'): 'U',
-    (104, 2, 'F', 'L'): 'U',
-    (104, 2, 'F', 'R'): 'U',
-    (104, 2, 'F', 'U'): 'U',
-    (104, 2, 'L', 'B'): 'D',
-    (104, 2, 'L', 'D'): 'U',
-    (104, 2, 'L', 'F'): 'D',
-    (104, 2, 'L', 'U'): 'U',
-    (104, 2, 'R', 'B'): 'D',
-    (104, 2, 'R', 'D'): 'U',
-    (104, 2, 'R', 'F'): 'D',
-    (104, 2, 'R', 'U'): 'U',
-    (104, 2, 'U', 'B'): 'D',
-    (104, 2, 'U', 'F'): 'D',
-    (104, 2, 'U', 'L'): 'D',
-    (104, 2, 'U', 'R'): 'D',
-    (106, 85, 'B', 'D'): 'U',
-    (106, 85, 'B', 'L'): 'U',
-    (106, 85, 'B', 'R'): 'U',
-    (106, 85, 'B', 'U'): 'U',
-    (106, 85, 'D', 'B'): 'D',
-    (106, 85, 'D', 'F'): 'D',
-    (106, 85, 'D', 'L'): 'D',
-    (106, 85, 'D', 'R'): 'D',
-    (106, 85, 'F', 'D'): 'U',
-    (106, 85, 'F', 'L'): 'U',
-    (106, 85, 'F', 'R'): 'U',
-    (106, 85, 'F', 'U'): 'U',
-    (106, 85, 'L', 'B'): 'D',
-    (106, 85, 'L', 'D'): 'U',
-    (106, 85, 'L', 'F'): 'D',
-    (106, 85, 'L', 'U'): 'U',
-    (106, 85, 'R', 'B'): 'D',
-    (106, 85, 'R', 'D'): 'U',
-    (106, 85, 'R', 'F'): 'D',
-    (106, 85, 'R', 'U'): 'U',
-    (106, 85, 'U', 'B'): 'D',
-    (106, 85, 'U', 'F'): 'D',
-    (106, 85, 'U', 'L'): 'D',
-    (106, 85, 'U', 'R'): 'D',
-    (110, 31, 'B', 'D'): 'D',
-    (110, 31, 'B', 'L'): 'D',
-    (110, 31, 'B', 'R'): 'D',
-    (110, 31, 'B', 'U'): 'D',
-    (110, 31, 'D', 'B'): 'U',
-    (110, 31, 'D', 'F'): 'U',
-    (110, 31, 'D', 'L'): 'U',
-    (110, 31, 'D', 'R'): 'U',
-    (110, 31, 'F', 'D'): 'D',
-    (110, 31, 'F', 'L'): 'D',
-    (110, 31, 'F', 'R'): 'D',
-    (110, 31, 'F', 'U'): 'D',
-    (110, 31, 'L', 'B'): 'U',
-    (110, 31, 'L', 'D'): 'D',
-    (110, 31, 'L', 'F'): 'U',
-    (110, 31, 'L', 'U'): 'D',
-    (110, 31, 'R', 'B'): 'U',
-    (110, 31, 'R', 'D'): 'D',
-    (110, 31, 'R', 'F'): 'U',
-    (110, 31, 'R', 'U'): 'D',
-    (110, 31, 'U', 'B'): 'U',
-    (110, 31, 'U', 'F'): 'U',
-    (110, 31, 'U', 'L'): 'U',
-    (110, 31, 'U', 'R'): 'U',
-    (111, 90, 'B', 'D'): 'U',
-    (111, 90, 'B', 'L'): 'U',
-    (111, 90, 'B', 'R'): 'U',
-    (111, 90, 'B', 'U'): 'U',
-    (111, 90, 'D', 'B'): 'D',
-    (111, 90, 'D', 'F'): 'D',
-    (111, 90, 'D', 'L'): 'D',
-    (111, 90, 'D', 'R'): 'D',
-    (111, 90, 'F', 'D'): 'U',
-    (111, 90, 'F', 'L'): 'U',
-    (111, 90, 'F', 'R'): 'U',
-    (111, 90, 'F', 'U'): 'U',
-    (111, 90, 'L', 'B'): 'D',
-    (111, 90, 'L', 'D'): 'U',
-    (111, 90, 'L', 'F'): 'D',
-    (111, 90, 'L', 'U'): 'U',
-    (111, 90, 'R', 'B'): 'D',
-    (111, 90, 'R', 'D'): 'U',
-    (111, 90, 'R', 'F'): 'D',
-    (111, 90, 'R', 'U'): 'U',
-    (111, 90, 'U', 'B'): 'D',
-    (111, 90, 'U', 'F'): 'D',
-    (111, 90, 'U', 'L'): 'D',
-    (111, 90, 'U', 'R'): 'D',
-    (115, 36, 'B', 'D'): 'U',
-    (115, 36, 'B', 'L'): 'U',
-    (115, 36, 'B', 'R'): 'U',
-    (115, 36, 'B', 'U'): 'U',
-    (115, 36, 'D', 'B'): 'D',
-    (115, 36, 'D', 'F'): 'D',
-    (115, 36, 'D', 'L'): 'D',
-    (115, 36, 'D', 'R'): 'D',
-    (115, 36, 'F', 'D'): 'U',
-    (115, 36, 'F', 'L'): 'U',
-    (115, 36, 'F', 'R'): 'U',
-    (115, 36, 'F', 'U'): 'U',
-    (115, 36, 'L', 'B'): 'D',
-    (115, 36, 'L', 'D'): 'U',
-    (115, 36, 'L', 'F'): 'D',
-    (115, 36, 'L', 'U'): 'U',
-    (115, 36, 'R', 'B'): 'D',
-    (115, 36, 'R', 'D'): 'U',
-    (115, 36, 'R', 'F'): 'D',
-    (115, 36, 'R', 'U'): 'U',
-    (115, 36, 'U', 'B'): 'D',
-    (115, 36, 'U', 'F'): 'D',
-    (115, 36, 'U', 'L'): 'D',
-    (115, 36, 'U', 'R'): 'D',
-    (116, 95, 'B', 'D'): 'D',
-    (116, 95, 'B', 'L'): 'D',
-    (116, 95, 'B', 'R'): 'D',
-    (116, 95, 'B', 'U'): 'D',
-    (116, 95, 'D', 'B'): 'U',
-    (116, 95, 'D', 'F'): 'U',
-    (116, 95, 'D', 'L'): 'U',
-    (116, 95, 'D', 'R'): 'U',
-    (116, 95, 'F', 'D'): 'D',
-    (116, 95, 'F', 'L'): 'D',
-    (116, 95, 'F', 'R'): 'D',
-    (116, 95, 'F', 'U'): 'D',
-    (116, 95, 'L', 'B'): 'U',
-    (116, 95, 'L', 'D'): 'D',
-    (116, 95, 'L', 'F'): 'U',
-    (116, 95, 'L', 'U'): 'D',
-    (116, 95, 'R', 'B'): 'U',
-    (116, 95, 'R', 'D'): 'D',
-    (116, 95, 'R', 'F'): 'U',
-    (116, 95, 'R', 'U'): 'D',
-    (116, 95, 'U', 'B'): 'U',
-    (116, 95, 'U', 'F'): 'U',
-    (116, 95, 'U', 'L'): 'U',
-    (116, 95, 'U', 'R'): 'U',
-    (120, 41, 'B', 'D'): 'U',
-    (120, 41, 'B', 'L'): 'U',
-    (120, 41, 'B', 'R'): 'U',
-    (120, 41, 'B', 'U'): 'U',
-    (120, 41, 'D', 'B'): 'D',
-    (120, 41, 'D', 'F'): 'D',
-    (120, 41, 'D', 'L'): 'D',
-    (120, 41, 'D', 'R'): 'D',
-    (120, 41, 'F', 'D'): 'U',
-    (120, 41, 'F', 'L'): 'U',
-    (120, 41, 'F', 'R'): 'U',
-    (120, 41, 'F', 'U'): 'U',
-    (120, 41, 'L', 'B'): 'D',
-    (120, 41, 'L', 'D'): 'U',
-    (120, 41, 'L', 'F'): 'D',
-    (120, 41, 'L', 'U'): 'U',
-    (120, 41, 'R', 'B'): 'D',
-    (120, 41, 'R', 'D'): 'U',
-    (120, 41, 'R', 'F'): 'D',
-    (120, 41, 'R', 'U'): 'U',
-    (120, 41, 'U', 'B'): 'D',
-    (120, 41, 'U', 'F'): 'D',
-    (120, 41, 'U', 'L'): 'D',
-    (120, 41, 'U', 'R'): 'D',
-    (122, 149, 'B', 'D'): 'U',
-    (122, 149, 'B', 'L'): 'U',
-    (122, 149, 'B', 'R'): 'U',
-    (122, 149, 'B', 'U'): 'U',
-    (122, 149, 'D', 'B'): 'D',
-    (122, 149, 'D', 'F'): 'D',
-    (122, 149, 'D', 'L'): 'D',
-    (122, 149, 'D', 'R'): 'D',
-    (122, 149, 'F', 'D'): 'U',
-    (122, 149, 'F', 'L'): 'U',
-    (122, 149, 'F', 'R'): 'U',
-    (122, 149, 'F', 'U'): 'U',
-    (122, 149, 'L', 'B'): 'D',
-    (122, 149, 'L', 'D'): 'U',
-    (122, 149, 'L', 'F'): 'D',
-    (122, 149, 'L', 'U'): 'U',
-    (122, 149, 'R', 'B'): 'D',
-    (122, 149, 'R', 'D'): 'U',
-    (122, 149, 'R', 'F'): 'D',
-    (122, 149, 'R', 'U'): 'U',
-    (122, 149, 'U', 'B'): 'D',
-    (122, 149, 'U', 'F'): 'D',
-    (122, 149, 'U', 'L'): 'D',
-    (122, 149, 'U', 'R'): 'D',
-    (123, 148, 'B', 'D'): 'U',
-    (123, 148, 'B', 'L'): 'U',
-    (123, 148, 'B', 'R'): 'U',
-    (123, 148, 'B', 'U'): 'U',
-    (123, 148, 'D', 'B'): 'D',
-    (123, 148, 'D', 'F'): 'D',
-    (123, 148, 'D', 'L'): 'D',
-    (123, 148, 'D', 'R'): 'D',
-    (123, 148, 'F', 'D'): 'U',
-    (123, 148, 'F', 'L'): 'U',
-    (123, 148, 'F', 'R'): 'U',
-    (123, 148, 'F', 'U'): 'U',
-    (123, 148, 'L', 'B'): 'D',
-    (123, 148, 'L', 'D'): 'U',
-    (123, 148, 'L', 'F'): 'D',
-    (123, 148, 'L', 'U'): 'U',
-    (123, 148, 'R', 'B'): 'D',
-    (123, 148, 'R', 'D'): 'U',
-    (123, 148, 'R', 'F'): 'D',
-    (123, 148, 'R', 'U'): 'U',
-    (123, 148, 'U', 'B'): 'D',
-    (123, 148, 'U', 'F'): 'D',
-    (123, 148, 'U', 'L'): 'D',
-    (123, 148, 'U', 'R'): 'D',
-    (124, 147, 'B', 'D'): 'D',
-    (124, 147, 'B', 'L'): 'D',
-    (124, 147, 'B', 'R'): 'D',
-    (124, 147, 'B', 'U'): 'D',
-    (124, 147, 'D', 'B'): 'U',
-    (124, 147, 'D', 'F'): 'U',
-    (124, 147, 'D', 'L'): 'U',
-    (124, 147, 'D', 'R'): 'U',
-    (124, 147, 'F', 'D'): 'D',
-    (124, 147, 'F', 'L'): 'D',
-    (124, 147, 'F', 'R'): 'D',
-    (124, 147, 'F', 'U'): 'D',
-    (124, 147, 'L', 'B'): 'U',
-    (124, 147, 'L', 'D'): 'D',
-    (124, 147, 'L', 'F'): 'U',
-    (124, 147, 'L', 'U'): 'D',
-    (124, 147, 'R', 'B'): 'U',
-    (124, 147, 'R', 'D'): 'D',
-    (124, 147, 'R', 'F'): 'U',
-    (124, 147, 'R', 'U'): 'D',
-    (124, 147, 'U', 'B'): 'U',
-    (124, 147, 'U', 'F'): 'U',
-    (124, 147, 'U', 'L'): 'U',
-    (124, 147, 'U', 'R'): 'U',
-    (127, 72, 'B', 'D'): 'D',
-    (127, 72, 'B', 'L'): 'D',
-    (127, 72, 'B', 'R'): 'D',
-    (127, 72, 'B', 'U'): 'D',
-    (127, 72, 'D', 'B'): 'U',
-    (127, 72, 'D', 'F'): 'U',
-    (127, 72, 'D', 'L'): 'U',
-    (127, 72, 'D', 'R'): 'U',
-    (127, 72, 'F', 'D'): 'D',
-    (127, 72, 'F', 'L'): 'D',
-    (127, 72, 'F', 'R'): 'D',
-    (127, 72, 'F', 'U'): 'D',
-    (127, 72, 'L', 'B'): 'U',
-    (127, 72, 'L', 'D'): 'D',
-    (127, 72, 'L', 'F'): 'U',
-    (127, 72, 'L', 'U'): 'D',
-    (127, 72, 'R', 'B'): 'U',
-    (127, 72, 'R', 'D'): 'D',
-    (127, 72, 'R', 'F'): 'U',
-    (127, 72, 'R', 'U'): 'D',
-    (127, 72, 'U', 'B'): 'U',
-    (127, 72, 'U', 'F'): 'U',
-    (127, 72, 'U', 'L'): 'U',
-    (127, 72, 'U', 'R'): 'U',
-    (128, 73, 'B', 'D'): 'D',
-    (128, 73, 'B', 'L'): 'D',
-    (128, 73, 'B', 'R'): 'D',
-    (128, 73, 'B', 'U'): 'D',
-    (128, 73, 'D', 'B'): 'U',
-    (128, 73, 'D', 'F'): 'U',
-    (128, 73, 'D', 'L'): 'U',
-    (128, 73, 'D', 'R'): 'U',
-    (128, 73, 'F', 'D'): 'D',
-    (128, 73, 'F', 'L'): 'D',
-    (128, 73, 'F', 'R'): 'D',
-    (128, 73, 'F', 'U'): 'D',
-    (128, 73, 'L', 'B'): 'U',
-    (128, 73, 'L', 'D'): 'D',
-    (128, 73, 'L', 'F'): 'U',
-    (128, 73, 'L', 'U'): 'D',
-    (128, 73, 'R', 'B'): 'U',
-    (128, 73, 'R', 'D'): 'D',
-    (128, 73, 'R', 'F'): 'U',
-    (128, 73, 'R', 'U'): 'D',
-    (128, 73, 'U', 'B'): 'U',
-    (128, 73, 'U', 'F'): 'U',
-    (128, 73, 'U', 'L'): 'U',
-    (128, 73, 'U', 'R'): 'U',
-    (129, 74, 'B', 'D'): 'U',
-    (129, 74, 'B', 'L'): 'U',
-    (129, 74, 'B', 'R'): 'U',
-    (129, 74, 'B', 'U'): 'U',
-    (129, 74, 'D', 'B'): 'D',
-    (129, 74, 'D', 'F'): 'D',
-    (129, 74, 'D', 'L'): 'D',
-    (129, 74, 'D', 'R'): 'D',
-    (129, 74, 'F', 'D'): 'U',
-    (129, 74, 'F', 'L'): 'U',
-    (129, 74, 'F', 'R'): 'U',
-    (129, 74, 'F', 'U'): 'U',
-    (129, 74, 'L', 'B'): 'D',
-    (129, 74, 'L', 'D'): 'U',
-    (129, 74, 'L', 'F'): 'D',
-    (129, 74, 'L', 'U'): 'U',
-    (129, 74, 'R', 'B'): 'D',
-    (129, 74, 'R', 'D'): 'U',
-    (129, 74, 'R', 'F'): 'D',
-    (129, 74, 'R', 'U'): 'U',
-    (129, 74, 'U', 'B'): 'D',
-    (129, 74, 'U', 'F'): 'D',
-    (129, 74, 'U', 'L'): 'D',
-    (129, 74, 'U', 'R'): 'D',
-    (131, 49, 'B', 'D'): 'U',
-    (131, 49, 'B', 'L'): 'U',
-    (131, 49, 'B', 'R'): 'U',
-    (131, 49, 'B', 'U'): 'U',
-    (131, 49, 'D', 'B'): 'D',
-    (131, 49, 'D', 'F'): 'D',
-    (131, 49, 'D', 'L'): 'D',
-    (131, 49, 'D', 'R'): 'D',
-    (131, 49, 'F', 'D'): 'U',
-    (131, 49, 'F', 'L'): 'U',
-    (131, 49, 'F', 'R'): 'U',
-    (131, 49, 'F', 'U'): 'U',
-    (131, 49, 'L', 'B'): 'D',
-    (131, 49, 'L', 'D'): 'U',
-    (131, 49, 'L', 'F'): 'D',
-    (131, 49, 'L', 'U'): 'U',
-    (131, 49, 'R', 'B'): 'D',
-    (131, 49, 'R', 'D'): 'U',
-    (131, 49, 'R', 'F'): 'D',
-    (131, 49, 'R', 'U'): 'U',
-    (131, 49, 'U', 'B'): 'D',
-    (131, 49, 'U', 'F'): 'D',
-    (131, 49, 'U', 'L'): 'D',
-    (131, 49, 'U', 'R'): 'D',
-    (135, 97, 'B', 'D'): 'D',
-    (135, 97, 'B', 'L'): 'D',
-    (135, 97, 'B', 'R'): 'D',
-    (135, 97, 'B', 'U'): 'D',
-    (135, 97, 'D', 'B'): 'U',
-    (135, 97, 'D', 'F'): 'U',
-    (135, 97, 'D', 'L'): 'U',
-    (135, 97, 'D', 'R'): 'U',
-    (135, 97, 'F', 'D'): 'D',
-    (135, 97, 'F', 'L'): 'D',
-    (135, 97, 'F', 'R'): 'D',
-    (135, 97, 'F', 'U'): 'D',
-    (135, 97, 'L', 'B'): 'U',
-    (135, 97, 'L', 'D'): 'D',
-    (135, 97, 'L', 'F'): 'U',
-    (135, 97, 'L', 'U'): 'D',
-    (135, 97, 'R', 'B'): 'U',
-    (135, 97, 'R', 'D'): 'D',
-    (135, 97, 'R', 'F'): 'U',
-    (135, 97, 'R', 'U'): 'D',
-    (135, 97, 'U', 'B'): 'U',
-    (135, 97, 'U', 'F'): 'U',
-    (135, 97, 'U', 'L'): 'U',
-    (135, 97, 'U', 'R'): 'U',
-    (136, 48, 'B', 'D'): 'D',
-    (136, 48, 'B', 'L'): 'D',
-    (136, 48, 'B', 'R'): 'D',
-    (136, 48, 'B', 'U'): 'D',
-    (136, 48, 'D', 'B'): 'U',
-    (136, 48, 'D', 'F'): 'U',
-    (136, 48, 'D', 'L'): 'U',
-    (136, 48, 'D', 'R'): 'U',
-    (136, 48, 'F', 'D'): 'D',
-    (136, 48, 'F', 'L'): 'D',
-    (136, 48, 'F', 'R'): 'D',
-    (136, 48, 'F', 'U'): 'D',
-    (136, 48, 'L', 'B'): 'U',
-    (136, 48, 'L', 'D'): 'D',
-    (136, 48, 'L', 'F'): 'U',
-    (136, 48, 'L', 'U'): 'D',
-    (136, 48, 'R', 'B'): 'U',
-    (136, 48, 'R', 'D'): 'D',
-    (136, 48, 'R', 'F'): 'U',
-    (136, 48, 'R', 'U'): 'D',
-    (136, 48, 'U', 'B'): 'U',
-    (136, 48, 'U', 'F'): 'U',
-    (136, 48, 'U', 'L'): 'U',
-    (136, 48, 'U', 'R'): 'U',
-    (140, 98, 'B', 'D'): 'D',
-    (140, 98, 'B', 'L'): 'D',
-    (140, 98, 'B', 'R'): 'D',
-    (140, 98, 'B', 'U'): 'D',
-    (140, 98, 'D', 'B'): 'U',
-    (140, 98, 'D', 'F'): 'U',
-    (140, 98, 'D', 'L'): 'U',
-    (140, 98, 'D', 'R'): 'U',
-    (140, 98, 'F', 'D'): 'D',
-    (140, 98, 'F', 'L'): 'D',
-    (140, 98, 'F', 'R'): 'D',
-    (140, 98, 'F', 'U'): 'D',
-    (140, 98, 'L', 'B'): 'U',
-    (140, 98, 'L', 'D'): 'D',
-    (140, 98, 'L', 'F'): 'U',
-    (140, 98, 'L', 'U'): 'D',
-    (140, 98, 'R', 'B'): 'U',
-    (140, 98, 'R', 'D'): 'D',
-    (140, 98, 'R', 'F'): 'U',
-    (140, 98, 'R', 'U'): 'D',
-    (140, 98, 'U', 'B'): 'U',
-    (140, 98, 'U', 'F'): 'U',
-    (140, 98, 'U', 'L'): 'U',
-    (140, 98, 'U', 'R'): 'U',
-    (141, 47, 'B', 'D'): 'D',
-    (141, 47, 'B', 'L'): 'D',
-    (141, 47, 'B', 'R'): 'D',
-    (141, 47, 'B', 'U'): 'D',
-    (141, 47, 'D', 'B'): 'U',
-    (141, 47, 'D', 'F'): 'U',
-    (141, 47, 'D', 'L'): 'U',
-    (141, 47, 'D', 'R'): 'U',
-    (141, 47, 'F', 'D'): 'D',
-    (141, 47, 'F', 'L'): 'D',
-    (141, 47, 'F', 'R'): 'D',
-    (141, 47, 'F', 'U'): 'D',
-    (141, 47, 'L', 'B'): 'U',
-    (141, 47, 'L', 'D'): 'D',
-    (141, 47, 'L', 'F'): 'U',
-    (141, 47, 'L', 'U'): 'D',
-    (141, 47, 'R', 'B'): 'U',
-    (141, 47, 'R', 'D'): 'D',
-    (141, 47, 'R', 'F'): 'U',
-    (141, 47, 'R', 'U'): 'D',
-    (141, 47, 'U', 'B'): 'U',
-    (141, 47, 'U', 'F'): 'U',
-    (141, 47, 'U', 'L'): 'U',
-    (141, 47, 'U', 'R'): 'U',
-    (145, 99, 'B', 'D'): 'U',
-    (145, 99, 'B', 'L'): 'U',
-    (145, 99, 'B', 'R'): 'U',
-    (145, 99, 'B', 'U'): 'U',
-    (145, 99, 'D', 'B'): 'D',
-    (145, 99, 'D', 'F'): 'D',
-    (145, 99, 'D', 'L'): 'D',
-    (145, 99, 'D', 'R'): 'D',
-    (145, 99, 'F', 'D'): 'U',
-    (145, 99, 'F', 'L'): 'U',
-    (145, 99, 'F', 'R'): 'U',
-    (145, 99, 'F', 'U'): 'U',
-    (145, 99, 'L', 'B'): 'D',
-    (145, 99, 'L', 'D'): 'U',
-    (145, 99, 'L', 'F'): 'D',
-    (145, 99, 'L', 'U'): 'U',
-    (145, 99, 'R', 'B'): 'D',
-    (145, 99, 'R', 'D'): 'U',
-    (145, 99, 'R', 'F'): 'D',
-    (145, 99, 'R', 'U'): 'U',
-    (145, 99, 'U', 'B'): 'D',
-    (145, 99, 'U', 'F'): 'D',
-    (145, 99, 'U', 'L'): 'D',
-    (145, 99, 'U', 'R'): 'D',
-    (147, 124, 'B', 'D'): 'U',
-    (147, 124, 'B', 'L'): 'U',
-    (147, 124, 'B', 'R'): 'U',
-    (147, 124, 'B', 'U'): 'U',
-    (147, 124, 'D', 'B'): 'D',
-    (147, 124, 'D', 'F'): 'D',
-    (147, 124, 'D', 'L'): 'D',
-    (147, 124, 'D', 'R'): 'D',
-    (147, 124, 'F', 'D'): 'U',
-    (147, 124, 'F', 'L'): 'U',
-    (147, 124, 'F', 'R'): 'U',
-    (147, 124, 'F', 'U'): 'U',
-    (147, 124, 'L', 'B'): 'D',
-    (147, 124, 'L', 'D'): 'U',
-    (147, 124, 'L', 'F'): 'D',
-    (147, 124, 'L', 'U'): 'U',
-    (147, 124, 'R', 'B'): 'D',
-    (147, 124, 'R', 'D'): 'U',
-    (147, 124, 'R', 'F'): 'D',
-    (147, 124, 'R', 'U'): 'U',
-    (147, 124, 'U', 'B'): 'D',
-    (147, 124, 'U', 'F'): 'D',
-    (147, 124, 'U', 'L'): 'D',
-    (147, 124, 'U', 'R'): 'D',
-    (148, 123, 'B', 'D'): 'D',
-    (148, 123, 'B', 'L'): 'D',
-    (148, 123, 'B', 'R'): 'D',
-    (148, 123, 'B', 'U'): 'D',
-    (148, 123, 'D', 'B'): 'U',
-    (148, 123, 'D', 'F'): 'U',
-    (148, 123, 'D', 'L'): 'U',
-    (148, 123, 'D', 'R'): 'U',
-    (148, 123, 'F', 'D'): 'D',
-    (148, 123, 'F', 'L'): 'D',
-    (148, 123, 'F', 'R'): 'D',
-    (148, 123, 'F', 'U'): 'D',
-    (148, 123, 'L', 'B'): 'U',
-    (148, 123, 'L', 'D'): 'D',
-    (148, 123, 'L', 'F'): 'U',
-    (148, 123, 'L', 'U'): 'D',
-    (148, 123, 'R', 'B'): 'U',
-    (148, 123, 'R', 'D'): 'D',
-    (148, 123, 'R', 'F'): 'U',
-    (148, 123, 'R', 'U'): 'D',
-    (148, 123, 'U', 'B'): 'U',
-    (148, 123, 'U', 'F'): 'U',
-    (148, 123, 'U', 'L'): 'U',
-    (148, 123, 'U', 'R'): 'U',
-    (149, 122, 'B', 'D'): 'D',
-    (149, 122, 'B', 'L'): 'D',
-    (149, 122, 'B', 'R'): 'D',
-    (149, 122, 'B', 'U'): 'D',
-    (149, 122, 'D', 'B'): 'U',
-    (149, 122, 'D', 'F'): 'U',
-    (149, 122, 'D', 'L'): 'U',
-    (149, 122, 'D', 'R'): 'U',
-    (149, 122, 'F', 'D'): 'D',
-    (149, 122, 'F', 'L'): 'D',
-    (149, 122, 'F', 'R'): 'D',
-    (149, 122, 'F', 'U'): 'D',
-    (149, 122, 'L', 'B'): 'U',
-    (149, 122, 'L', 'D'): 'D',
-    (149, 122, 'L', 'F'): 'U',
-    (149, 122, 'L', 'U'): 'D',
-    (149, 122, 'R', 'B'): 'U',
-    (149, 122, 'R', 'D'): 'D',
-    (149, 122, 'R', 'F'): 'U',
-    (149, 122, 'R', 'U'): 'D',
-    (149, 122, 'U', 'B'): 'U',
-    (149, 122, 'U', 'F'): 'U',
-    (149, 122, 'U', 'L'): 'U',
-    (149, 122, 'U', 'R'): 'U',
-}
 
 
 swaps_555 = {'2B': (0, 1, 2, 3, 4, 5, 79, 84, 89, 94, 99, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 10, 28, 29, 30, 31, 9, 33, 34, 35, 36, 8, 38, 39, 40, 41, 7, 43, 44, 45, 46, 6, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 145, 80, 81, 82, 83, 144, 85, 86, 87, 88, 143, 90, 91, 92, 93, 142, 95, 96, 97, 98, 141, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 27, 32, 37, 42, 47, 146, 147, 148, 149, 150),
